% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{AerodynamicResistance}
\alias{AerodynamicResistance}
\title{Aerodynamic Resistance. Eq 4.2.25 in David R Maidment, Handbook of Hydrology
\code{AerodynamicResistance}}
\usage{
AerodynamicResistance(Uz, hc, Z_u, Z_e, VON_KARMAN = 0.4)
}
\arguments{
\item{Uz}{Wind speed at height z in m s^-1.}

\item{hc}{Height of crop in m.}

\item{Z_u}{Height of wind measurements in m.}

\item{Z_e}{Height of humidity measurements in m.}

\item{VON_KARMAN}{von Karman's constant. Default is 0.41.}
}
\value{
Aerodynamic resistance in s m^-1.
}
\description{
Aerodynamic Resistance. Eq 4.2.25 in David R Maidment, Handbook of Hydrology
\code{AerodynamicResistance}
}
\examples{
hc = cbind(0.1, 1, 10)
uz = 5
zu = 2
ze = 2
AerodynamicResistance(uz, hc, zu, ze)
}
