% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelBC.R
\name{BC}
\alias{BC}
\title{Generate the default Boundary Condition
\code{BC}}
\usage{
BC(x = cbind(0:100, 0), start = as.POSIXct("2000-01-01"))
}
\arguments{
\item{x}{matrix of BC. c(days, value)}

\item{start}{Start time, POSIXct.}
}
\value{
Time-series data of BC
}
\description{
Generate the default Boundary Condition
\code{BC}
}
