% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{BulkSurfaceResistance}
\alias{BulkSurfaceResistance}
\title{Bulk Surface Resistance, Eq 4.2.22 in David R Maidment, Handbook of Hydrology
\code{BulkSurfaceResistance}}
\usage{
BulkSurfaceResistance(lai)
}
\arguments{
\item{lai}{Leaf Area Index [m2 m-2]}
}
\value{
Bulk Surface Resistance [s m-1]
}
\description{
Bulk Surface Resistance, Eq 4.2.22 in David R Maidment, Handbook of Hydrology
\code{BulkSurfaceResistance}
}
