% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{CanopyResistance}
\alias{CanopyResistance}
\title{Canopy resistance
\code{CanopyResistance}}
\usage{
CanopyResistance(rmin, lai)
}
\arguments{
\item{rmin}{Minimum resistance in s m^-1.}

\item{lai}{Leaf Area Index in m2 m^-2.}
}
\value{
Canopy resistance in s m^-1.
}
\description{
Canopy resistance
\code{CanopyResistance}
}
