% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{CommonPoints}
\alias{CommonPoints}
\title{Find the common coordinates in two groups of coordinates.}
\usage{
CommonPoints(x, y)
}
\arguments{
\item{x, y}{Matrix of coodinates.}
}
\value{
Matrix (m*2). Columns are index of x rows and y rows.
}
\description{
Find the common coordinates in two groups of coordinates.
}
