% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{correctRiverSlope}
\alias{correctRiverSlope}
\title{Convert the bed slope of river, to avoid negative/zero slope
\code{correctRiverSlope}}
\usage{
correctRiverSlope(pr, minSlope = 1e-05, maxrun = 500)
}
\arguments{
\item{pr}{SHUD.RIVER class}

\item{minSlope}{Minimum slope}

\item{maxrun}{Maximum number to run the loops}
}
\value{
SpatialLinesDataFrame
}
\description{
Convert the bed slope of river, to avoid negative/zero slope
\code{correctRiverSlope}
}
