% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{crs_to_string}
\alias{crs_to_string}
\title{Convert CRS to string representation}
\usage{
crs_to_string(crs)
}
\arguments{
\item{crs}{CRS object (various types)}
}
\value{
Character string
}
\description{
Convert CRS to string representation
}
\keyword{internal}
