% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFilter.R
\name{datafilter.riv}
\alias{datafilter.riv}
\title{Filter the river data with a filter
\code{datafilter.riv}}
\usage{
datafilter.riv(x, filter = NULL, plot = TRUE)
}
\arguments{
\item{x}{Input data}

\item{filter}{Data filter}

\item{plot}{Whether plot the data}
}
\value{
Matrix information, c('ID','Vmin','Vmax', 'Filter')
}
\description{
Filter the river data with a filter
\code{datafilter.riv}
}
