% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{days_in_year}
\alias{days_in_year}
\title{Number of days in years.}
\usage{
days_in_year(years)
}
\arguments{
\item{years}{Years}
}
\value{
Number of days between first and last years.
}
\description{
Number of days in years.
}
\examples{
days_in_year(1980)
days_in_year(1983:2019)
}
