% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBalance.R
\name{DeltaS}
\alias{DeltaS}
\title{Calculate the Change of Storage.
\code{DeltaS}}
\usage{
DeltaS(x, x0 = x[t1, ], t1 = 1, t2 = nrow(x))
}
\arguments{
\item{x}{Time-Series Matrix}

\item{x0}{Intial condition or the values of first time-step}

\item{t1}{Time-step one, default = 1}

\item{t2}{Time-step two, default = nrow(x)}
}
\value{
A vector
}
\description{
Calculate the Change of Storage.
\code{DeltaS}
}
