% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelClasses.R
\docType{data}
\name{.shud}
\alias{.shud}
\title{This is .shud environment.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.shud
}
\description{
This is .shud environment.
}
\keyword{datasets}
