% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{Eudist}
\alias{Eudist}
\title{Eudilic Distance
\code{Eudist}}
\usage{
Eudist(p1, p2)
}
\arguments{
\item{p1}{coordinates of point 1}

\item{p2}{coordinates of point 2}
}
\value{
Distance between \code{p1} and \code{p2}
}
\description{
Eudilic Distance
\code{Eudist}
}
\examples{
p1 = c(1,1)
p2 = c(2,2)
Eudist(p1,p2)
}
