% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCalibration.R
\name{EXEC}
\alias{EXEC}
\title{Execute the program in terminal.
\code{EXEC}}
\usage{
EXEC(CV, CMD.EXE, calibfile, outpath, fn.log)
}
\arguments{
\item{CV}{Calibration variables, list.}

\item{CMD.EXE}{Command.}

\item{calibfile}{Filepath to calibration parameters.}

\item{outpath}{Path to output data.}

\item{fn.log}{Path of log file saving program stdout.}
}
\description{
Execute the program in terminal.
\code{EXEC}
}
