% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{extract_epsg}
\alias{extract_epsg}
\title{Extract EPSG code from CRS string}
\usage{
extract_epsg(crs_str)
}
\arguments{
\item{crs_str}{Character string with CRS definition}
}
\value{
Integer EPSG code or NA
}
\description{
Extract EPSG code from CRS string
}
\keyword{internal}
