% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydro_obs.R
\name{fdc}
\alias{fdc}
\title{Default Comparison of the observation and simulation
\code{fdc}}
\usage{
fdc(val, xlab = "Exceedance (\%)", ylab = "value")
}
\arguments{
\item{val}{Matrix or data.frame, number of column=2}
}
\value{
ggplot handler
}
\description{
Default Comparison of the observation and simulation
\code{fdc}
}
\examples{
obs=rnorm(100)
sim=obs+rnorm(100)/2
fdc(cbind(obs, sim))
}
