% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{filebackup}
\alias{filebackup}
\title{Write a backup file if the file exist
\code{filebackup}}
\usage{
filebackup(file, backup = TRUE)
}
\arguments{
\item{file}{file name}

\item{backup}{if backup, TRUE/FALSE}
}
\description{
Write a backup file if the file exist
\code{filebackup}
}
