% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{ForcingCoverage}
\alias{ForcingCoverage}
\title{Generate the coverage map for forcing sites.
\code{ForcingCoverage}}
\usage{
ForcingCoverage(
  sp.meteoSite = NULL,
  filenames = paste0(sp.meteoSite$ID, ".csv"),
  pcs,
  gcs = sf::st_crs(4326),
  dem,
  wbd,
  enlarge = 10000
)
}
\arguments{
\item{sp.meteoSite}{sf POINT object or SpatialPoints*}

\item{filenames}{Output filenames associated with the forcing sites.}

\item{pcs}{Projected Coordinate System (crs object)}

\item{gcs}{Geographic Coordinate System (crs object)}

\item{dem}{DEM raster (SpatRaster or RasterLayer)}

\item{wbd}{watershed boundary (sf or SpatialPolygons*)}

\item{enlarge}{enlarge factor for the boundary.}
}
\value{
sf object
}
\description{
Generate the coverage map for forcing sites.
\code{ForcingCoverage}
}
