% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_landcover.R
\name{fun.lairl}
\alias{fun.lairl}
\title{Generate the default landcover parameters for NLCD classification
\code{fun.lairl}}
\usage{
fun.lairl(lc, years = 2000)
}
\arguments{
\item{lc}{land classe codes in NLCD classes.}

\item{years}{numeric years.}
}
\value{
Default fun.lairl parameters, a list $LAI, $RL
}
\description{
Generate the default landcover parameters for NLCD classification
\code{fun.lairl}
}
\examples{
lc = c(43, 23, 81, 11)
lr=fun.lairl(lc, years=2000:2001)
par(mfrow=c(2,1))
col=1:length(lc)
plot(lr$LAI, col=col, main='LAI');
legend('top', paste0(lc), col=col, lwd=1)
plot(lr$RL, col=col, main='Roughness Length');
legend('top', paste0(lc), col=col, lwd=1)
}
