% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_output.R
\name{get_default_variables}
\alias{get_default_variables}
\title{Get default output variables}
\usage{
get_default_variables()
}
\value{
Character vector of variable names
}
\description{
Returns a vector of commonly used SHUD output variables.
}
\seealso{
Other output-io: 
\code{\link{load_output_data}()},
\code{\link{read_output_header}()},
\code{\link{read_output}()}
}
\concept{output-io}
