% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_output.R
\name{get_model_version}
\alias{get_model_version}
\title{Get model version from environment}
\usage{
get_model_version()
}
\value{
Numeric model version
}
\description{
Get model version from environment
}
\keyword{internal}
