% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_netcdf.R
\name{get_nc_info}
\alias{get_nc_info}
\title{Get NetCDF file information}
\usage{
get_nc_info(ncid)
}
\arguments{
\item{ncid}{NetCDF connection object}
}
\value{
List with file information
}
\description{
Extracts basic information about NetCDF file structure.
}
\seealso{
Other netcdf-io: 
\code{\link{nc_to_raster}()},
\code{\link{open_nc_file}()},
\code{\link{read_nc_data}()},
\code{\link{read_nc_time}()}
}
\concept{netcdf-io}
