% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_output.R
\name{get_output_path}
\alias{get_output_path}
\title{Get output path from environment}
\usage{
get_output_path()
}
\value{
Character string with output path
}
\description{
Get output path from environment
}
\keyword{internal}
