% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_output.R
\name{get_project_name}
\alias{get_project_name}
\title{Get project name from environment}
\usage{
get_project_name()
}
\value{
Character string with project name
}
\description{
Get project name from environment
}
\keyword{internal}
