% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getAquiferDepth}
\alias{getAquiferDepth}
\title{Calculate the AquiferDepth of mesh cells
\code{getAquiferDepth}}
\usage{
getAquiferDepth(pm = readmesh())
}
\arguments{
\item{pm}{\code{shud.mesh}}
}
\value{
aquifer thickness
}
\description{
Calculate the AquiferDepth of mesh cells
\code{getAquiferDepth}
}
