% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getArea}
\alias{getArea}
\title{Calculate the area of the cells
\code{getArea}}
\usage{
getArea(pm = readmesh())
}
\arguments{
\item{pm}{\code{shud.mesh}}
}
\value{
Area of cells
}
\description{
Calculate the area of the cells
\code{getArea}
}
