% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getElevation}
\alias{getElevation}
\title{Calculate the elevation of mesh cells
\code{getElevation}}
\usage{
getElevation(pm = readmesh())
}
\arguments{
\item{pm}{\code{shud.mesh}}
}
\value{
elevation
}
\description{
Calculate the elevation of mesh cells
\code{getElevation}
}
