% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getLakeEleID}
\alias{getLakeEleID}
\title{Get the Lake Element Index
\code{getLakeEleID}}
\usage{
getLakeEleID(pa = readatt())
}
\arguments{
\item{pa}{\code{shud.att}}
}
\value{
index of lake element index
}
\description{
Get the Lake Element Index
\code{getLakeEleID}
}
