% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{getOutlets}
\alias{getOutlets}
\title{Get outlet ID(s)
\code{getOutlets}}
\usage{
getOutlets(pr = readriv())
}
\arguments{
\item{pr}{SHUD.RIVER class}
}
\value{
Index of outlets, numeric
}
\description{
Get outlet ID(s)
\code{getOutlets}
}
