% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getRiverNodes}
\alias{getRiverNodes}
\title{Get the From/To nodes of the river.
\code{getRiverNodes}}
\usage{
getRiverNodes(
  spr = sf::st_read(file.path(shud.filein()["inpath"], "gis", "river.shp"), quiet = TRUE)
)
}
\arguments{
\item{spr}{\code{sf} LINESTRING object for river streams; legacy spatial
input is accepted for compatibility.}
}
\value{
a list, c(points, FT_ID)
}
\description{
Get the From/To nodes of the river.
\code{getRiverNodes}
}
