% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getVertex}
\alias{getVertex}
\title{Calculate the Vertex of the cells
\code{getVertex}}
\usage{
getVertex(pm = readmesh())
}
\arguments{
\item{pm}{\code{shud.mesh}}
}
\value{
Vertex of \code{shud.mesh}, dim = c(ncell, vertex, 4), 3-dimension = c('x','y','AqD', 'zmax')
}
\description{
Calculate the Vertex of the cells
\code{getVertex}
}
