% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{hydrograph}
\alias{hydrograph}
\title{Deprecated wrapper for \code{hydrograph()}}
\usage{
hydrograph(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{plot_hydrograph}}.}
}
\value{
A plot object from \code{\link{plot_hydrograph}}.
}
\description{
Legacy compatibility wrapper for \code{\link{plot_hydrograph}}.
}
