% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{LAI2hc}
\alias{LAI2hc}
\title{Calculate vegetation height from Leaf Area Index (LAI), Eq 4.2.23 in David R Maidment, Handbook of Hydrology
\code{LAI2hc}}
\usage{
LAI2hc(lai)
}
\arguments{
\item{lai}{Leaf Area Index (LAI) [m2 m-2]}
}
\value{
Height of crop [m]
}
\description{
Calculate vegetation height from Leaf Area Index (LAI), Eq 4.2.23 in David R Maidment, Handbook of Hydrology
\code{LAI2hc}
}
\examples{
lai = seq(0, 10, length.out=100)
hc = LAI2hc(lai)
# lai=5.5+log(hc)*1.5
plot(lai, hc)
}
