% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_landcover.R
\name{LaiRf.GLC}
\alias{LaiRf.GLC}
\title{Generate the default LAI and Roughness Length for USGS GLC
\code{LaiRf.GLC}}
\usage{
LaiRf.GLC(lc = NULL, years = 2000 + 1:2, if.daily = FALSE)
}
\arguments{
\item{lc}{land classe codes}

\item{years}{numeric years.}

\item{if.daily}{whether daily output}
}
\value{
Default LaiRf.GLC parameters, a list $LAI, $RL
}
\description{
Generate the default LAI and Roughness Length for USGS GLC
\code{LaiRf.GLC}
}
\examples{
lc = 1:14
lr=LaiRf.GLC(lc, years=2000:2001)
par(mfrow=c(2,1))
col=1:length(lc)
plot(lr$LAI, col=col, main='LAI');
legend('top', paste0(lc), col=col, lwd=1)
plot(lr$RL, col=col, main='Roughness Length');
legend('top', paste0(lc), col=col, lwd=1)
}
