% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{LatentHeat}
\alias{LatentHeat}
\title{Latent Heat, Eq 4.2.1 in David R Maidment, Handbook of Hydrology
\code{shud.ic}}
\usage{
LatentHeat(Temp)
}
\arguments{
\item{Temp}{Air temperature, [C]}
}
\value{
Latent heat, lambda in [MJ/kg]
}
\description{
Latent Heat, Eq 4.2.1 in David R Maidment, Handbook of Hydrology
\code{shud.ic}
}
