% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTF.R
\name{lc.default}
\alias{lc.default}
\title{The default land cover parameters from University of Marryland Classification
\code{lc.default}}
\usage{
lc.default()
}
\value{
Default land cover parameters
}
\description{
The default land cover parameters from University of Marryland Classification
\code{lc.default}
}
