% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_landcover.R
\name{lc.GLC}
\alias{lc.GLC}
\title{The default land cover parameters from USGS GLC
\code{lc.GLC}}
\usage{
lc.GLC()
}
\value{
Default land cover parameters
}
\description{
The default land cover parameters from USGS GLC
\code{lc.GLC}
}
