% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_landcover.R
\name{lc.NLCD}
\alias{lc.NLCD}
\title{Generate the default land cover parameters of NLCD classes.
\code{PTF.NLCD}}
\usage{
lc.NLCD(lc)
}
\arguments{
\item{lc}{NLCD (2001 and later version) land use code.}
}
\value{
Default land cover parameters of NLCD classes.
}
\description{
Generate the default land cover parameters of NLCD classes.
\code{PTF.NLCD}
}
