% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_landcover.R
\name{lc.UMD}
\alias{lc.UMD}
\title{The default land cover parameters from UMD
\code{lc.UMD}}
\usage{
lc.UMD()
}
\value{
Default land cover parameters
}
\description{
The default land cover parameters from UMD
\code{lc.UMD}
}
