% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_landcover.R
\name{lc_EQ}
\alias{lc_EQ}
\title{convert equation from Univeristy of Marryland classes to NLCD classes
\code{lc_EQ}}
\usage{
lc_EQ(dtab, lc)
}
\arguments{
\item{dtab}{conversion table.}

\item{lc}{land classe codes in NLCD classes.}
}
\value{
Converted values.
}
\description{
convert equation from Univeristy of Marryland classes to NLCD classes
\code{lc_EQ}
}
\references{
Lele Shu(2017), Gopal Bhatt(2012)
}
