% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydro_obs.R
\name{LineFit}
\alias{LineFit}
\title{Linefit the observation and simulation
\code{LineFit}}
\usage{
LineFit(x, y = NULL, xlab = "Observation", ylab = "Simulation", ...)
}
\arguments{
\item{x}{Observation data or matrix/data.frame consist of observation and simulation}

\item{y}{Simulation data}

\item{xlab}{xlab, default is 'Observation'}

\item{ylab}{ylab, default is 'Simulation'}

\item{...}{more options for stat_poly_eq()}
}
\value{
Triangle mesh of model domain.
}
\description{
Linefit the observation and simulation
\code{LineFit}
}
\examples{
obs <- rnorm(100)
sim <- obs+rnorm(100)/2
p <- LineFit(cbind(obs,sim))
p
if(require(ggpmisc)){
formula = y ~ x
p+ ggpmisc::stat_poly_eq(aes(label = paste(..eq.label.., ..rr.label.., sep = "~~~")), 
label.x.npc = "right", label.y.npc = 0.15,
formula = formula, parse = TRUE, size = 3)
}
}
