% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTF.R
\name{MeltFactor}
\alias{MeltFactor}
\title{Generate the default melt factor
\code{MeltFactor}}
\usage{
MeltFactor(years)
}
\arguments{
\item{years}{Years.}
}
\value{
Default default melt factor.
}
\description{
Generate the default melt factor
\code{MeltFactor}
}
\examples{
years = 2000:2001
x = MeltFactor(years=years)
plot(x)
}
