% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfo.R
\name{MeshAtt}
\alias{MeshAtt}
\title{Summary of the attributions on Mesh
\code{MeshAtt}}
\usage{
MeshAtt(
  pm = read_mesh(),
  att = read_att(),
  soil = read_soil(),
  geol = read_geol(),
  lc = read_lc()
)
}
\arguments{
\item{pm}{SHUD.MESH object}

\item{att}{SHUD attribute table}

\item{soil}{.para.soil of input}

\item{geol}{.para.geol of input}

\item{lc}{.para.lc of input}
}
\value{
Attributes of each element.
}
\description{
Summary of the attributions on Mesh
\code{MeshAtt}
}
