% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R, R/gis_core.R
\name{MeshData2Raster}
\alias{MeshData2Raster}
\title{SpatialData to Raster
\code{MeshData2Raster}}
\usage{
MeshData2Raster(...)

MeshData2Raster(...)
}
\arguments{
\item{...}{Arguments passed to \code{mesh_to_raster}}

\item{x}{vector or matrix, length/nrow is number of cells.}

\item{rmask}{mask of the mesh file}

\item{stack}{Whether export the stack, only when the x is a matrix, i.e. (Ntime x Ncell).}

\item{proj}{Projejction parameter}

\item{pm}{shud mesh}

\item{method}{method for interpolation, default = 'idw'}

\item{plot}{Whether plot the result.}
}
\value{
Raster map
}
\description{
This function is deprecated. Use \code{\link{mesh_to_raster}} instead.
}
\keyword{internal}
