% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sinks.R
\name{meshSinks}
\alias{meshSinks}
\title{Get the Sink cells in SHUD mesh
\code{meshSinks}}
\usage{
meshSinks(pm = readmesh(), details = FALSE)
}
\arguments{
\item{pm}{SHUD mesh}

\item{details}{Whether return details of sinks infomration.}
}
\value{
Vector of TRUE/FALSE or data.frame
}
\description{
Get the Sink cells in SHUD mesh
\code{meshSinks}
}
