% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfo.R
\name{ModelInfo}
\alias{ModelInfo}
\title{Summary of the basic information of model input data.
\code{ModelInfo}}
\usage{
ModelInfo(
  path = shud.filein()["outpath"],
  spr = readriv.sp(),
  crs.pcs = raster::crs(spr)
)
}
\arguments{
\item{path}{Where to export the basic model infomation.}

\item{crs.pcs}{The crs4string that defines the (x,y) in the model.}
}
\value{
Basic model infomation, figures and tables
}
\description{
Summary of the basic information of model input data.
\code{ModelInfo}
}
