% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLCD.R
\name{NLCD.colors}
\alias{NLCD.colors}
\title{Return NLCD colors.
\code{NLCD.colors}}
\source{
See the link of NLCD: \url{http://www.mrlc.gov/nlcd11_leg.php}
}
\usage{
NLCD.colors(x, def.col = "#EEEEEE")
}
\arguments{
\item{x, }{which is the integer number defined in NLCD 2001,2006, and 2011.  Any value out of (0,100) will be assigned as black.}

\item{def.col}{Default color, if the landuse code is missing.}
}
\value{
HEX color strings
}
\description{
Return NLCD colors.
\code{NLCD.colors}
}
\keyword{Landuse,}
\keyword{NLCD,}
\keyword{colormap}
