% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLCD.R
\name{NLCD.names}
\alias{NLCD.names}
\title{Return NLCD class names.
\code{NLCD.names}}
\usage{
NLCD.names(x, def.name = "n/a")
}
\arguments{
\item{x, }{which is the integer number defined in NLCD 2001,2006, and 2011.  Any value out of (0,100) will be assigned as black.}

\item{def.name}{Default name.}
}
\value{
Characters
}
\description{
Return NLCD class names.
\code{NLCD.names}
}
\keyword{Landuse}
\keyword{NLCD,}
