% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{NodeIDList}
\alias{NodeIDList}
\title{Return node indices for line geometry
\code{NodeIDList}}
\usage{
NodeIDList(sp, coord = get_coords(sp))
}
\arguments{
\item{sp}{\code{sf} LINESTRING object; legacy \code{SpatialLines*} is accepted
for compatibility.}

\item{coord}{Coordinate of vertex in \code{sp}}
}
\value{
list of node indices for each line feature.
}
\description{
Return node indices for line geometry
\code{NodeIDList}
}
