% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_netcdf.R
\name{open_nc_file}
\alias{open_nc_file}
\title{Open NetCDF file with error handling}
\usage{
open_nc_file(file)
}
\arguments{
\item{file}{Character. Path to NetCDF file}
}
\value{
NetCDF connection object
}
\description{
Wrapper around ncdf4::nc_open with improved error handling.
}
\seealso{
Other netcdf-io: 
\code{\link{get_nc_info}()},
\code{\link{nc_to_raster}()},
\code{\link{read_nc_data}()},
\code{\link{read_nc_time}()}
}
\concept{netcdf-io}
