% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{PET_PM}
\alias{PET_PM}
\title{Potential Evapotranspiration with Pennmann-Monteith equation.
\code{PET_PM}}
\usage{
PET_PM(
  Wind,
  Temp,
  RH,
  RadNet,
  Press = NULL,
  WindHeight = 10,
  Veg_height = 0.12,
  albedo = 0.23,
  Res_surf = 70,
  Elevation = NULL
)
}
\arguments{
\item{Wind}{Windspeed in m s^-1.}

\item{Temp}{Air temperature in C.}

\item{RH}{Relative humidity as a fraction from 0 to 1.}

\item{RadNet}{Net solar radiation in W m^-2.}

\item{Press}{Pressure in kPa. If \code{NULL}, pressure is computed from
\code{Elevation}.}

\item{WindHeight}{Height that wind was measured in m.}

\item{Veg_height}{Height of vegetation in m.}

\item{albedo}{Albedo as a fraction.}

\item{Res_surf}{Aerodynamic surface resistance in s m^-1.}

\item{Elevation}{Elevation in m.}
}
\value{
Potential evapotranspiration in mm day^-1.
}
\description{
Potential Evapotranspiration with Pennmann-Monteith equation.
\code{PET_PM}
}
