% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plot_animate}
\alias{plot_animate}
\title{Plot animation maps
\code{plot.animate}}
\usage{
plot_animate(x, id = NULL, nmap = 10, rlist = NULL)
}
\arguments{
\item{x}{Time-series data.}

\item{id}{Which row(s) to plot in the animation.}

\item{nmap}{Number of maps to be plot.}

\item{rlist}{RasterStack of the maps.}
}
\description{
Plot animation maps
\code{plot.animate}
}
