% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial.R
\name{plot_spatial}
\alias{plot_spatial}
\title{Spatial Plotting Functions}
\description{
Functions for visualizing spatial data including mesh domains, rasters,
and vector data using modern spatial libraries (terra and sf).
}
\keyword{internal}
