% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{png.control}
\alias{png.control}
\title{Deprecated wrapper for \code{png.control()}}
\usage{
png.control(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[grDevices]{png}}.}
}
\value{
The result from \code{\link[grDevices]{png}}.
}
\description{
Legacy compatibility wrapper for \code{grDevices::png()}.
}
