% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{polygonArea}
\alias{polygonArea}
\title{Calculate the Polygon Area;}
\usage{
polygonArea(X, Y)
}
\arguments{
\item{X}{required}

\item{Y}{required}
}
\value{
Area
}
\description{
Calculate the Polygon Area;
}
