% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCalibration.R
\name{Pop2Calib}
\alias{Pop2Calib}
\title{Apply population to calibration values.
\code{Pop2Calib}}
\usage{
Pop2Calib(x, calib, range)
}
\arguments{
\item{x}{Population data between 0 and 1.}

\item{calib}{calib data.}

\item{range}{Range of the calib value.}
}
\description{
Apply population to calibration values.
\code{Pop2Calib}
}
