% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelCalibration.R
\name{pre.files}
\alias{pre.files}
\title{Generate file path for calibration populations.
\code{pre.files}}
\usage{
pre.files(iGen, pop, CV)
}
\arguments{
\item{iGen}{Index of generations}

\item{pop}{Population of the generation}

\item{CV}{Calibration variables, list.}
}
\description{
Generate file path for calibration populations.
\code{pre.files}
}
