% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{PressureElevation}
\alias{PressureElevation}
\title{Air pressure at the elevation, Allen(1998) Eq(7)
\code{PressureElevation}}
\usage{
PressureElevation(z)
}
\arguments{
\item{z}{Elevation [m]}
}
\value{
Air pressure at the elevation [kPa]
}
\description{
Air pressure at the elevation, Allen(1998) Eq(7)
\code{PressureElevation}
}
