% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{ProjectCoordinate}
\alias{ProjectCoordinate}
\title{Re-project coordinates betwen GCS and PCS
\code{ProjectCoordinate}}
\usage{
ProjectCoordinate(x, proj4string, P2G = TRUE)
}
\arguments{
\item{x}{2-column matrix of coordinates.}

\item{proj4string}{proj4string}

\item{P2G}{if TRUE, a cartographic projection into lat/long, otherwise projects from lat/long into a cartographic projection.}
}
\value{
Basic model infomation, figures and tables
}
\description{
Re-project coordinates betwen GCS and PCS
\code{ProjectCoordinate}
}
