% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_Projection.R
\name{projection}
\alias{projection}
\title{Coordinate Projection Functions}
\description{
Functions for creating and working with coordinate reference systems (CRS)
using modern spatial libraries (sf).
}
