% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTF.R
\name{PTF}
\alias{PTF}
\title{Pedotransfer function to generate soil/geol parameter from soil texture
\code{PTF}}
\usage{
PTF(x = t(matrix(c(33, 33, 2, 1.4), ncol = 5, nrow = 4)),
  topsoil = TRUE)
}
\arguments{
\item{x}{data.frame or matrix, column = c(Silt_perc, Clay_perc, OrganicMatter _perc, BulkDensity g/cm3)}

\item{topsoil}{TRUE/FALSE, default = TRUE}
}
\value{
Hydraulic parameters, matrix
}
\description{
Pedotransfer function to generate soil/geol parameter from soil texture
\code{PTF}
}
\examples{
x=cbind(10,40, 1:20, 1.5)
y=PTF(x)
apply(y[,-1], 2, summary)
plot(x[,3], y[,2])
}
