% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTF.R
\name{PTF.soil}
\alias{PTF.soil}
\title{Pedotransfer function to generate soil parameter
\code{PTF.soil}}
\usage{
PTF.soil(x = t(matrix(c(33, 33, 2, 1.4), ncol = 5, nrow = 4)),
  topsoil = TRUE, rm.outlier = FALSE)
}
\arguments{
\item{x}{data.frame or matrix, column = c(Silt_perc, Clay_perc, OrganicMatter _perc, BulkDensity g/cm3)}

\item{topsoil}{TRUE/FALSE, default = TRUE}

\item{rm.outlier}{Whether replace the outliers in each column with the mean value of the column.}
}
\value{
Hydraulic parameters, matrix
}
\description{
Pedotransfer function to generate soil parameter
\code{PTF.soil}
}
\examples{
x=cbind(10,40, 1:20, 1.5)
y=PTF.soil(x)
apply(y[,-1], 2, summary)
plot(x[,3], y[,2])
}
