% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hydro_obs.R
\name{QvsO}
\alias{QvsO}
\title{Default Comparison of the observation and simulation
\code{QvsO}}
\usage{
QvsO(qq, sim = qq[, 2], obs = qq[, 1], ...)
}
\arguments{
\item{qq}{Matrix or data.frame. Column 1 is observation, while colum 2 is simulation.}

\item{sim}{Simulation time-seris data;}

\item{obs}{Observation time-seris data;}

\item{...}{more options for plot()}
}
\description{
Default Comparison of the observation and simulation
\code{QvsO}
}
\examples{
obs=rnorm(100)
sim=obs+rnorm(100)/2
QvsO(cbind(obs,sim))
}
