% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{read.auto}
\alias{read.auto}
\title{Prepare configure file for autoSHUD
\code{read.auto}}
\usage{
read.auto(fn)
}
\value{
Character of full path of input files for SHUD model
}
\description{
Prepare configure file for autoSHUD
\code{read.auto}
}
