% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{read.df}
\alias{read.df}
\title{Read the matrix or data.frame file which has (nrow, ncol) in header
\code{read.df}}
\usage{
read.df(file, text = readLines(file))
}
\arguments{
\item{file}{full path of file}

\item{text}{text return from readLines(file)}
}
\value{
a list of matrix
}
\description{
Read the matrix or data.frame file which has (nrow, ncol) in header
\code{read.df}
}
