% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTSD.R
\name{read.tsd}
\alias{read.tsd}
\title{Read the time-series file, which has c(nrow, ncol, timestring) in header
\code{read.tsd}}
\usage{
read.tsd(file)
}
\arguments{
\item{file}{full path of file}
}
\value{
xts
}
\description{
Read the time-series file, which has c(nrow, ncol, timestring) in header
\code{read.tsd}
}
