% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_att}
\alias{read_att}
\alias{readatt}
\title{Read SHUD attributes file}
\usage{
read_att(file = shud.filein()["md.att"])

readatt(file = shud.filein()["md.att"])
}
\arguments{
\item{file}{Character. Full path to the .att file}
}
\value{
Matrix containing attribute data
}
\description{
Reads a SHUD attributes file (.att) and returns attribute data.
}
\section{Deprecated}{

\code{readatt()} is deprecated. Use \code{read_att()} instead.
}

\examples{
\dontrun{
att <- read_att("model.att")
}
}
\seealso{
Other shud-io: 
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river_sp}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
