% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_calib}
\alias{read_calib}
\alias{readcalib}
\title{Read SHUD calibration file}
\usage{
read_calib(file = shud.filein()["md.calib"])

readcalib(file = shud.filein()["md.calib"])
}
\arguments{
\item{file}{Character. Full path to the .calib file}
}
\value{
Data frame containing calibration data
}
\description{
Reads a SHUD calibration file (.calib) and returns calibration data.
}
\section{Deprecated}{

\code{readcalib()} is deprecated. Use \code{read_calib()} instead.
}

\examples{
\dontrun{
calib <- read_calib("model.calib")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
