% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_config}
\alias{read_config}
\alias{readconfig}
\title{Read SHUD configuration file}
\usage{
read_config(file = shud.filein()["md.para"])

readconfig(file = shud.filein()["md.para"])
}
\arguments{
\item{file}{Character. Full path to the configuration file}
}
\value{
Data frame containing configuration data
}
\description{
Reads a SHUD configuration file (.para or .calib) and returns configuration data.
}
\section{Deprecated}{

\code{readconfig()} is deprecated. Use \code{read_config()} instead.
}

\examples{
\dontrun{
config <- read_config("model.para")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river_sp}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
