% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_df}
\alias{read_df}
\alias{read.df}
\title{Read matrix or data frame file with header}
\usage{
read_df(file, text = readLines(file), sep = "\\t")

read.df(file, text = readLines(file), sep = "\\t")
}
\arguments{
\item{file}{Character. Full path to file}

\item{text}{Character vector. Text content from readLines(file). If provided, file is ignored}

\item{sep}{Character. Separator for the table (default: tab)}
}
\value{
List of matrices/data frames
}
\description{
Reads a file containing matrix or data frame data with (nrow, ncol) in header.
This is a core utility function used by other SHUD file readers.
}
\section{Deprecated}{

\code{read.df()} is deprecated. Use \code{read_df()} instead.
}

\examples{
\dontrun{
data <- read_df("data.txt")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
