% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_forc_csv}
\alias{read_forc_csv}
\alias{readforc.csv}
\title{Read SHUD forcing CSV files}
\usage{
read_forc_csv(file = shud.filein()["md.forc"], id = NULL)

readforc.csv(file = shud.filein()["md.forc"], id = NULL)
}
\arguments{
\item{file}{Character. Full path to the .forc file}

\item{id}{Numeric vector. Index of forcing sites to read. If NULL, returns average of all sites}
}
\value{
Forcing data as xts object or list of xts objects
}
\description{
Reads the CSV files referenced in a SHUD forcing file (.forc).
}
\section{Deprecated}{

\code{readforc.csv()} is deprecated. Use \code{read_forc_csv()} instead.
}

\examples{
\dontrun{
forc_data <- read_forc_csv("model.forc", id = 1:3)
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river_sp}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
