% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_geol}
\alias{read_geol}
\alias{readgeol}
\title{Read SHUD geology file}
\usage{
read_geol(file = shud.filein()["md.geol"])

readgeol(file = shud.filein()["md.geol"])
}
\arguments{
\item{file}{Character. Full path to the .geol file}
}
\value{
Matrix containing geology data
}
\description{
Reads a SHUD geology file (.geol) and returns geology data.
}
\section{Deprecated}{

\code{readgeol()} is deprecated. Use \code{read_geol()} instead.
}

\examples{
\dontrun{
geol <- read_geol("model.geol")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river_sp}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
