% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_ic}
\alias{read_ic}
\alias{readic}
\title{Read SHUD initial conditions file}
\usage{
read_ic(file = shud.filein()["md.ic"])

readic(file = shud.filein()["md.ic"])
}
\arguments{
\item{file}{Character. Full path to the .ic file}
}
\value{
List containing initial condition data
}
\description{
Reads a SHUD initial conditions file (.ic) and returns initial condition data.
}
\section{Deprecated}{

\code{readic()} is deprecated. Use \code{read_ic()} instead.
}

\examples{
\dontrun{
ic <- read_ic("model.ic")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
