% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_timeseries.R
\name{read_lai}
\alias{read_lai}
\alias{readlai}
\title{Read LAI time series file}
\usage{
read_lai(file = shud.filein()["ts.lai"])

readlai(file = shud.filein()["ts.lai"])
}
\arguments{
\item{file}{Character. Full path to LAI time series file}
}
\value{
List of time series data with LAI and RL (Roughness Length) components
}
\description{
Reads a LAI (Leaf Area Index) time series file (.ts.lai) and returns
properly named time series components.
}
\section{Deprecated}{

\code{readlai()} is deprecated. Use \code{read_lai()} instead.
}

\examples{
\dontrun{
lai_data <- read_lai("model.ts.lai")
plot(lai_data$LAI)
plot(lai_data$RL)
}
}
\seealso{
Other timeseries-io: 
\code{\link{read_tsd}()},
\code{\link{ts_to_daily}()},
\code{\link{ts_to_df}()},
\code{\link{write_tsd}()},
\code{\link{write_xts}()}
}
\concept{timeseries-io}
