% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_lc}
\alias{read_lc}
\alias{readlc}
\title{Read SHUD land cover file}
\usage{
read_lc(file = shud.filein()["md.lc"])

readlc(file = shud.filein()["md.lc"])
}
\arguments{
\item{file}{Character. Full path to the .lc file}
}
\value{
Matrix containing land cover data
}
\description{
Reads a SHUD land cover file (.lc) and returns land cover data.
}
\section{Deprecated}{

\code{readlc()} is deprecated. Use \code{read_lc()} instead.
}

\examples{
\dontrun{
lc <- read_lc("model.lc")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river_sp}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
