% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_mesh}
\alias{read_mesh}
\alias{readmesh}
\title{Read SHUD mesh file}
\usage{
read_mesh(file = shud.filein()["md.mesh"])

readmesh(file = shud.filein()["md.mesh"])
}
\arguments{
\item{file}{Character. Full path to the .mesh file}
}
\value{
SHUD.MESH object containing mesh and point data
}
\description{
Reads a SHUD mesh file (.mesh) and returns a SHUD.MESH object.
}
\section{Deprecated}{

\code{readmesh()} is deprecated. Use \code{read_mesh()} instead.
}

\examples{
\dontrun{
mesh <- read_mesh("model.mesh")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
