% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_netcdf.R
\name{read_nc_time}
\alias{read_nc_time}
\title{Read NetCDF time dimension}
\usage{
read_nc_time(ncid)
}
\arguments{
\item{ncid}{NetCDF connection object from ncdf4::nc_open()}
}
\value{
POSIXct vector of time stamps
}
\description{
Extracts time information from NetCDF files and converts to POSIXct objects.
Supports various time formats and units commonly used in climate data.
}
\examples{
\dontrun{
library(ncdf4)
nc <- nc_open("climate_data.nc")
times <- read_nc_time(nc)
nc_close(nc)
}
}
\seealso{
Other netcdf-io: 
\code{\link{get_nc_info}()},
\code{\link{nc_to_raster}()},
\code{\link{open_nc_file}()},
\code{\link{read_nc_data}()}
}
\concept{netcdf-io}
