% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_output.R
\name{read_output_header}
\alias{read_output_header}
\title{Read output file header information}
\usage{
read_output_header(keyword = NULL, file = NULL, path = get_output_path())
}
\arguments{
\item{keyword}{Character. Keyword of the output file}

\item{file}{Character. Full path to output file (optional)}

\item{path}{Character. Path to output directory}
}
\value{
Character string containing header information
}
\description{
Reads header information from SHUD output files for inspection.
}
\examples{
\dontrun{
header_info <- read_output_header("eleygw")
}
}
\seealso{
Other output-io: 
\code{\link{get_default_variables}()},
\code{\link{load_output_data}()},
\code{\link{read_output}()}
}
\concept{output-io}
