% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_para}
\alias{read_para}
\alias{readpara}
\title{Read SHUD parameters file}
\usage{
read_para(file = shud.filein()["md.para"])

readpara(file = shud.filein()["md.para"])
}
\arguments{
\item{file}{Character. Full path to the .para file}
}
\value{
Data frame containing parameter data
}
\description{
Reads a SHUD parameters file (.para) and returns parameter data.
}
\section{Deprecated}{

\code{readpara()} is deprecated. Use \code{read_para()} instead.
}

\examples{
\dontrun{
para <- read_para("model.para")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_river_sp}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
