% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_river}
\alias{read_river}
\alias{readriv}
\title{Read SHUD river file}
\usage{
read_river(file = shud.filein()["md.riv"])

readriv(file = shud.filein()["md.riv"])
}
\arguments{
\item{file}{Character. Full path to the .riv file}
}
\value{
SHUD.RIVER object containing river network data
}
\description{
Reads a SHUD river file (.riv) and returns a SHUD.RIVER object.
}
\section{Deprecated}{

\code{readriv()} is deprecated. Use \code{read_river()} instead.
}

\examples{
\dontrun{
river <- read_river("model.riv")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
