% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_river_sp}
\alias{read_river_sp}
\alias{readriv.sp}
\title{Read river shapefile}
\usage{
read_river_sp(file = file.path(shud.filein()["inpath"], "gis", "river.shp"))

readriv.sp(file = file.path(shud.filein()["inpath"], "gis", "river.shp"))
}
\arguments{
\item{file}{Character. Full path to river shapefile}
}
\value{
SpatialLines object
}
\description{
Reads a river shapefile and returns a SpatialLines object.
Note: This function uses legacy spatial libraries and will be updated
in future versions to use sf.
}
\section{Deprecated}{

\code{readriv.sp()} is deprecated. Use \code{read_river_sp()} instead.
}

\examples{
\dontrun{
river_sp <- read_river_sp("rivers.shp")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
