% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_rivseg}
\alias{read_rivseg}
\alias{readrivseg}
\title{Read SHUD river segments file}
\usage{
read_rivseg(file = shud.filein()["md.rivseg"])

readrivseg(file = shud.filein()["md.rivseg"])
}
\arguments{
\item{file}{Character. Full path to the .rivseg file}
}
\value{
Matrix containing river segment data
}
\description{
Reads a SHUD river segments file (.rivseg) and returns segment data.
}
\section{Deprecated}{

\code{readrivseg()} is deprecated. Use \code{read_rivseg()} instead.
}

\examples{
\dontrun{
rivseg <- read_rivseg("model.rivseg")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
