% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_soil}
\alias{read_soil}
\alias{readsoil}
\title{Read SHUD soil file}
\usage{
read_soil(file = shud.filein()["md.soil"])

readsoil(file = shud.filein()["md.soil"])
}
\arguments{
\item{file}{Character. Full path to the .soil file}
}
\value{
Matrix containing soil data
}
\description{
Reads a SHUD soil file (.soil) and returns soil data.
}
\section{Deprecated}{

\code{readsoil()} is deprecated. Use \code{read_soil()} instead.
}

\examples{
\dontrun{
soil <- read_soil("model.soil")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river_sp}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
