% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_timeseries.R
\name{read_tsd}
\alias{read_tsd}
\alias{read.tsd}
\title{Read time series data file}
\usage{
read_tsd(file)

read.tsd(file)
}
\arguments{
\item{file}{Character. Full path to time series file}
}
\value{
List of xts objects containing time series data
}
\description{
Reads a time series file with header containing (nrow, ncol, timestring).
This function handles SHUD-specific time series format with automatic
time parsing and xts object creation.
}
\section{Deprecated}{

\code{read.tsd()} is deprecated. Use \code{read_tsd()} instead.
}

\examples{
\dontrun{
ts_data <- read_tsd("timeseries.tsd")
lai_data <- read_lai("model.ts.lai")
}
}
\seealso{
Other timeseries-io: 
\code{\link{read_lai}()},
\code{\link{ts_to_daily}()},
\code{\link{ts_to_df}()},
\code{\link{write_tsd}()},
\code{\link{write_xts}()}
}
\concept{timeseries-io}
