% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readatt}
\alias{readatt}
\title{Read the .att file
\code{readmesh}}
\usage{
readatt(file = shud.filein()["md.att"])
}
\arguments{
\item{file}{full path of file}
}
\value{
matrix
}
\description{
Read the .att file
\code{readmesh}
}
