% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readcalib}
\alias{readcalib}
\title{Read the .calib file
\code{readcalib}}
\usage{
readcalib(file = shud.filein()["md.calib"])
}
\arguments{
\item{file}{full path of file}
}
\value{
.calib
}
\description{
Read the .calib file
\code{readcalib}
}
