% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readchannel}
\alias{readchannel}
\title{Read the .rivchn file
\code{readchannel}}
\usage{
readchannel(file = shud.filein()["md.rivseg"])
}
\arguments{
\item{file}{full path of .rivchn file}
}
\value{
matrix
}
\description{
Read the .rivchn file
\code{readchannel}
}
