% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readconfig}
\alias{readconfig}
\title{Read the configuration (.para, .calib) file
\code{readconfig}}
\usage{
readconfig(file = shud.filein()["md.para"])
}
\arguments{
\item{file}{full path of file}
}
\value{
.para or .calib
}
\description{
Read the configuration (.para, .calib) file
\code{readconfig}
}
