% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readforc.csv}
\alias{readforc.csv}
\title{Read the .csv files in .forc file
\code{readforc.csv}}
\usage{
readforc.csv(file = shud.filein()["md.forc"], id = NULL)
}
\arguments{
\item{file}{full path of file}

\item{id}{Index of the forcing sites.  default = NULL, which return average values of all sites.}
}
\value{
forcing data, list.
}
\description{
Read the .csv files in .forc file
\code{readforc.csv}
}
