% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readforc.fn}
\alias{readforc.fn}
\title{Read the .forc file
\code{readlc}}
\usage{
readforc.fn(file = shud.filein()["md.forc"])
}
\arguments{
\item{file}{full path of file}
}
\value{
data.frame of forcing sites.
}
\description{
Read the .forc file
\code{readlc}
}
