% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readgeol}
\alias{readgeol}
\title{Read the .geol file
\code{readgeol}}
\usage{
readgeol(file = shud.filein()["md.geol"])
}
\arguments{
\item{file}{full path of file}
}
\value{
.geol
}
\description{
Read the .geol file
\code{readgeol}
}
