% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readic}
\alias{readic}
\title{Read the .ic file
\code{readic}}
\usage{
readic(file = shud.filein()["md.ic"])
}
\arguments{
\item{file}{full path of file}
}
\value{
.ic
}
\description{
Read the .ic file
\code{readic}
}
