% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTSD.R
\name{readlai}
\alias{readlai}
\title{Read the .ts.lai Time-Series file
\code{readlai}}
\usage{
readlai(file = shud.filein()["ts.lai"])
}
\arguments{
\item{file}{full path of file}
}
\value{
list of Time-Series data
}
\description{
Read the .ts.lai Time-Series file
\code{readlai}
}
