% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readlc}
\alias{readlc}
\title{Read the .lc file
\code{readlc}}
\usage{
readlc(file = shud.filein()["md.lc"])
}
\arguments{
\item{file}{full path of file}
}
\value{
.lc
}
\description{
Read the .lc file
\code{readlc}
}
