% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readmesh}
\alias{readmesh}
\title{Read the .mesh file
\code{readmesh}}
\usage{
readmesh(file = shud.filein()["md.mesh"])
}
\arguments{
\item{file}{full path of .mesh file}
}
\value{
Mesh domain
}
\description{
Read the .mesh file
\code{readmesh}
}
