% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF.R
\name{readnc}
\alias{readnc}
\title{Read the time tage from NC file.}
\usage{
readnc(ncid, varid = NULL, ext = NULL)
}
\arguments{
\item{ncid}{the ncid from nc_open()}

\item{varid}{The varid.}

\item{ext}{box coordinates of the subset.}
}
\value{
a POSIXct date-time objects
}
\description{
Read the time tage from NC file.
}
