% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF.R
\name{readnc.time}
\alias{readnc.time}
\title{Read the time tage from NC file.}
\usage{
readnc.time(ncid)
}
\arguments{
\item{ncid}{the ncid from nc_open()}
}
\value{
a POSIXct date-time objects
}
\description{
Read the time tage from NC file.
}
