% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readout.R
\name{readout.header}
\alias{readout.header}
\title{Read header information of output file}
\usage{
readout.header(
  keyword,
  path = get("outpath", envir = .shud),
  file = file.path(path, paste0(get("PRJNAME", envir = .shud), ".", keyword, ".dat"))
)
}
\arguments{
\item{keyword}{keyword of the output file. File name format is: projectname.keyword.dat}

\item{path}{path of the output file}

\item{file}{Full path of output file.}
}
\value{
Header information
}
\description{
Read header information of output file
}
\keyword{Could}
\keyword{and}
\keyword{be}
\keyword{for}
\keyword{format.}
\keyword{mesh}
\keyword{output.}
\keyword{read}
\keyword{reading}
\keyword{river}
\keyword{used}
