% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readpara}
\alias{readpara}
\title{Read the .para file
\code{readpara}}
\usage{
readpara(file = shud.filein()["md.para"])
}
\arguments{
\item{file}{full path of file}
}
\value{
.para
}
\description{
Read the .para file
\code{readpara}
}
