% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readriv.sp}
\alias{readriv.sp}
\title{Read the river shapefile
\code{readriv.sp}}
\usage{
readriv.sp(file = file.path(shud.filein()["inpath"], "gis", "river.shp"))
}
\arguments{
\item{file}{full path of spatialLines file}
}
\value{
spatialLines file
}
\description{
Read the river shapefile
\code{readriv.sp}
}
