% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readrivseg}
\alias{readrivseg}
\title{Read the .rivseg file
\code{readrivseg}}
\usage{
readrivseg(file = shud.filein()["md.rivseg"])
}
\arguments{
\item{file}{full path of .rivseg file}
}
\value{
matrix
}
\description{
Read the .rivseg file
\code{readrivseg}
}
