% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readinput.R
\name{readsoil}
\alias{readsoil}
\title{Read the .soil file
\code{readsoil}}
\usage{
readsoil(file = shud.filein()["md.soil"])
}
\arguments{
\item{file}{full path of file}
}
\value{
.soil
}
\description{
Read the .soil file
\code{readsoil}
}
