% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{removeholes}
\alias{removeholes}
\title{Deprecated wrapper for \code{removeholes()}}
\usage{
removeholes(sp, ...)
}
\description{
Uses modern terra/sf polygon handling to fill interior holes.
}
