% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfo.R
\name{RiverAtt}
\alias{RiverAtt}
\title{Summary of the attributions on Rivers
\code{RiverAtt}}
\usage{
RiverAtt(riv = readriv())
}
\arguments{
\item{river}{SHUD.RIVER}
}
\value{
Attributes of each river reach
}
\description{
Summary of the attributions on Rivers
\code{RiverAtt}
}
