% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{RiverLength}
\alias{RiverLength}
\title{Calculate the river length
\code{RiverLength}}
\usage{
RiverLength(pr)
}
\arguments{
\item{pr}{SHUD.RIVER class}
}
\value{
River length, vector
}
\description{
Calculate the river length
\code{RiverLength}
}
