% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{RiverSlope}
\alias{RiverSlope}
\title{Calculate the river slope
\code{RiverSlope}}
\usage{
RiverSlope(pr)
}
\arguments{
\item{pr}{SHUD.RIVER class}
}
\value{
River slopes, vector
}
\description{
Calculate the river slope
\code{RiverSlope}
}
