% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{RiverType}
\alias{RiverType}
\title{parameters for river types
\code{RiverType}}
\usage{
RiverType(n, width = 2 * (1:n))
}
\arguments{
\item{n}{number of types}
}
\value{
data.frame of parameters
}
\description{
parameters for river types
\code{RiverType}
}
