% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{SharedPoints}
\alias{SharedPoints}
\title{Find Shared points in SpatialLine*
\code{SharedPoints}}
\usage{
SharedPoints(sp, plot = FALSE)
}
\arguments{
\item{sp}{SpatialLines}

\item{plot}{Whether to plot the results for debugging}
}
\value{
Topologic relationship between components of SpatialLines.
}
\description{
Find Shared points in SpatialLine*
\code{SharedPoints}
}
