% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConfigure.R
\name{shud.calib}
\alias{shud.calib}
\title{Generate the default model calibration
\code{shud.calib}}
\usage{
shud.calib()
}
\value{
Default calibration values for model
}
\description{
Generate the default model calibration
\code{shud.calib}
}
