% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHUD_Env.R
\name{shud.env}
\alias{shud.env}
\title{Setup .shud environment.
\code{shud.env}}
\usage{
shud.env(prjname, inpath = file.path("input", prjname),
  outpath = file.path("output", paste0(prjname, ".out")),
  anapath = file.path(outpath, "SHUDtb"))
}
\arguments{
\item{prjname}{charactor of SHUD model project name.}

\item{inpath}{path of SHUD model input directory}

\item{outpath}{path of SHUD model output directory}

\item{anapath}{charactor of SHUD model project name.}
}
\description{
Setup .shud environment.
\code{shud.env}
}
