% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Project.R
\name{shud.filein}
\alias{shud.filein}
\title{Prepare Input file of SHUD model
\code{shud.filein}}
\usage{
shud.filein(projname = get("PRJNAME", envir = .shud),
  inpath = get("inpath", envir = .shud), outpath = get("outpath", envir
  = .shud))
}
\arguments{
\item{projname}{Character, project name, default= PRJNAME which is a global variable.}

\item{inpath}{SHUD model input directory, default = inpath which is global variable}

\item{outpath}{SHUD model output directory, default = outpath which is global variable}
}
\value{
Character of full path of input files for SHUD model
}
\description{
Prepare Input file of SHUD model
\code{shud.filein}
}
