% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConfigure.R
\name{shud.ic}
\alias{shud.ic}
\title{Generate SHUD initial condition tables
\code{shud.ic}}
\usage{
shud.ic(
  ncell,
  nriv,
  AqD = 10,
  stage = 0.1,
  p1 = 0.4,
  p2 = p1,
  lakestage = NULL
)
}
\arguments{
\item{ncell}{Number of triangles in SHUD.MESH domain}

\item{nriv}{Number of river lines}

\item{AqD}{Default aquifer depth}

\item{stage}{Default river stage}

\item{p1}{Fraction of unsaturated zone, from 0 to 1.}

\item{p2}{Fraction of groundwater, from 0 to 1.}

\item{lakestage}{Lake Stages. Null=No lake.}
}
\value{
List of initial condition
}
\description{
Generate SHUD initial condition tables
\code{shud.ic}
}
