% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{shud.mask}
\alias{shud.mask}
\title{Generate the terra raster mask of a mesh domain
\code{shud.mask}}
\usage{
shud.mask(pm = read_mesh(), proj = NULL, rr = NULL, n = 10000, cellsize = NULL)
}
\arguments{
\item{pm}{\code{shud.mesh}}

\item{proj}{Projection parameter}

\item{rr}{Default mask in .shud environment}

\item{n}{Number of grid}

\item{cellsize}{Resolution, defaul = NULL, resolution = extent / ngrids}
}
\value{
\code{terra::SpatRaster} mask
}
\description{
Generate the terra raster mask of a mesh domain
\code{shud.mask}
}
