% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelConfigure.R
\name{shud.para}
\alias{shud.para}
\title{Generate the default model configuration
\code{shud.para}}
\usage{
shud.para(nday = 10)
}
\arguments{
\item{nday}{Simulation periods (days)}
}
\value{
model configureation, Vector
}
\description{
Generate the default model configuration
\code{shud.para}
}
