% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_SimplifyPolylines.R
\name{SimplifybyLen}
\alias{SimplifybyLen}
\title{Deprecated: split legacy polylines by length
\code{SimplifybyLen}}
\source{
https://stackoverflow.com/questions/38700246/how-do-i-split-divide-polyline-shapefiles-into-equally-length-smaller-segments
}
\usage{
SimplifybyLen(sp, split_length = 20, plot.results = FALSE)
}
\arguments{
\item{sp}{\code{sf} line object; legacy \code{SpatialLines} input is accepted
for compatibility.}

\item{split_length}{the length of the segments to split the lines into, in
map units.}

\item{plot.results}{TRUE/FALSE}
}
\value{
Legacy \code{SpatialLinesDataFrame} for compatibility.
}
\description{
Deprecated: split legacy polylines by length
\code{SimplifybyLen}
}
