% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{SinglePolygon}
\alias{SinglePolygon}
\title{Convert MULTIPOLYGON geometry to single POLYGON features.}
\usage{
SinglePolygon(x, id = 0)
}
\arguments{
\item{x}{\code{sf} polygon object; legacy spatial input is accepted for
compatibility.}

\item{id}{Index of the sorted (decreasing) polygons to return. default = 0;}
}
\description{
Convert MULTIPOLYGON geometry to single POLYGON features.
}
