% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{sp.RiverDown}
\alias{sp.RiverDown}
\title{Deprecated: determine downstream index for legacy river lines
\code{sp.RiverDown}}
\usage{
sp.RiverDown(sp, coord = get_coords(sp))
}
\arguments{
\item{sp}{Legacy \code{SpatialLines*} or \code{sf} LINESTRING object.}

\item{coord}{coordinates of \code{sp}.}
}
\value{
Index of downstream for each segements
}
\description{
Deprecated: determine downstream index for legacy river lines
\code{sp.RiverDown}
}
\keyword{deprecated}
