% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{sp.RiverOrder}
\alias{sp.RiverOrder}
\title{Deprecated: calculate river order for legacy river lines
\code{sp.RiverOrder}}
\usage{
sp.RiverOrder(sp, coord = get_coords(sp))
}
\arguments{
\item{sp}{Legacy \code{SpatialLines*} or \code{sf} LINESTRING object.}

\item{coord}{coordinates of \code{sp}.}
}
\value{
Stream order vector.
}
\description{
Deprecated: calculate river order for legacy river lines
\code{sp.RiverOrder}
}
\examples{
\dontrun{
if (requireNamespace("sp", quietly = TRUE)) {
  library(sp)
  data(sac)
  riv=sac$riv
  ord = sp.RiverOrder(riv)
  print(ord)
  idx = sort(unique(ord))
  plot(riv, col=ord)
  legend('topleft', legend=idx, col=idx, lwd=1, lty=1)
}
}
}
\keyword{deprecated}
