% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_SimplifybyLen.R
\name{sp.simplifyLen}
\alias{sp.simplifyLen}
\title{Deprecated: simplify legacy polylines by length
\code{sp.simplifyLen}}
\usage{
sp.simplifyLen(sp, tol, threshold = tol/2)
}
\arguments{
\item{sp}{Legacy \code{SpatialLines} or \code{SpatialPolygons} object.}

\item{tol}{Tolerance}

\item{threshold}{Minimum length for a segment}
}
\description{
Deprecated: simplify legacy polylines by length
\code{sp.simplifyLen}
}
\keyword{deprecated}
