% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R, R/gis_core.R
\name{sp2raster}
\alias{sp2raster}
\title{SpatialData to Raster
\code{sp2raster}}
\usage{
sp2raster(...)

sp2raster(...)
}
\arguments{
\item{...}{Arguments passed to \code{vector_to_raster}}

\item{sp}{SpatialPolygon}

\item{mask}{Raster mask of mesh domain.}

\item{ngrids}{Number of grid along x direction.}

\item{resolution}{Resolution, defaul = NULL, resolution = extent / ngrids}

\item{field}{Index of field}
}
\value{
Raster map
}
\description{
This function is deprecated. Use \code{\link{vector_to_raster}} instead.
}
\keyword{internal}
