% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{toCentroid}
\alias{toCentroid}
\title{Find the centroid of the triangles
\code{toCentroid}}
\usage{
toCentroid(tri, point)
}
\arguments{
\item{tri}{Triangle, N x 3 dimension}

\item{point}{Coordinates.}
}
\value{
Coordinates of triangles centroids
}
\description{
Find the centroid of the triangles
\code{toCentroid}
}
