% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{triTopology}
\alias{triTopology}
\title{Determine the topological relation among triangles.}
\usage{
triTopology(tri)
}
\arguments{
\item{tri}{triangle defination, m x 3}
}
\value{
topological relation of triangles. m x 4; cols = c(ID, Nabor1, Nabor2, Nabor3)
}
\description{
Determine the topological relation among triangles.
}
