% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{ts2Daily}
\alias{ts2Daily}
\title{Convert the time-series data to daily data.}
\usage{
ts2Daily(x, FUN = mean, ...)
}
\arguments{
\item{x}{xts,zoo data.}

\item{FUN}{function.}

\item{...}{more options in xts::apply.daily()}
}
\value{
xts data whose time index is daily.
}
\description{
Convert the time-series data to daily data.
}
\examples{
library(xts)
x=as.xts(rnorm(100), order.by = as.POSIXct('2000-01-01')+ 1:100 * 3600*24)
xd = ts2Daily(x)
class(time(x))
class(time(xd))
}
