% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBalance.R
\name{ts2df}
\alias{ts2df}
\title{Convert Time-Series data to data.frame
\code{ts2df}}
\usage{
ts2df(x)
}
\arguments{
\item{x}{Time-Series data.}
}
\value{
data.frame. First column is Time.
}
\description{
Convert Time-Series data to data.frame
\code{ts2df}
}
