% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{ts2map}
\alias{ts2map}
\title{Plot map of time-series data
\code{ts2map}}
\usage{
ts2map(x, fun = mean, raster = TRUE, ...)
}
\arguments{
\item{x}{Time-series data}

\item{fun}{Functions to summarize the TS data.}

\item{raster}{Whether to make output as Raster}

\item{...}{More options in fun}
}
\description{
Plot map of time-series data
\code{ts2map}
}
