% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_timeseries.R
\name{ts_to_daily}
\alias{ts_to_daily}
\alias{ts2Daily}
\title{Convert time series to daily aggregation}
\usage{
ts_to_daily(x, FUN = mean)

ts2Daily(x, FUN = mean)
}
\arguments{
\item{x}{xts object with time series data}

\item{FUN}{Function to use for aggregation (default: mean)}
}
\value{
xts object with daily aggregated data
}
\description{
Aggregates time series data to daily values using specified function.
}
\section{Deprecated}{

\code{ts2Daily()} is deprecated. Use \code{ts_to_daily()} instead.
}

\examples{
\dontrun{
# Create hourly data
hourly_data <- xts::xts(runif(24*7), 
                       order.by = seq(Sys.time(), by = "hour", length.out = 24*7))

# Convert to daily means
daily_data <- ts_to_daily(hourly_data)
}
}
\seealso{
Other timeseries-io: 
\code{\link{read_lai}()},
\code{\link{read_tsd}()},
\code{\link{ts_to_df}()},
\code{\link{write_tsd}()},
\code{\link{write_xts}()}
}
\concept{timeseries-io}
