% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_timeseries.R
\name{ts_to_df}
\alias{ts_to_df}
\alias{ts2df}
\title{Convert time series to data frame}
\usage{
ts_to_df(x, time_col = "Time")

ts2df(x, time_col = "Time")
}
\arguments{
\item{x}{xts object with time series data}

\item{time_col}{Character. Name for time column (default: "Time")}
}
\value{
data.frame with time and data columns
}
\description{
Converts xts time series to data frame with time column.
}
\section{Deprecated}{

\code{ts2df()} is deprecated. Use \code{ts_to_df()} instead.
}

\examples{
\dontrun{
dates <- seq(as.Date("2020-01-01"), as.Date("2020-01-10"), by = "day")
ts_data <- xts::xts(runif(length(dates)), order.by = dates)
df_data <- ts_to_df(ts_data)
}
}
\seealso{
Other timeseries-io: 
\code{\link{read_lai}()},
\code{\link{read_tsd}()},
\code{\link{ts_to_daily}()},
\code{\link{write_tsd}()},
\code{\link{write_xts}()}
}
\concept{timeseries-io}
