% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_landcover.R
\name{tsLaiRf}
\alias{tsLaiRf}
\title{Generate the default LAI and Roughness Length for UMD classification system
\code{tsLaiRf}}
\usage{
tsLaiRf(type = 1)
}
\arguments{
\item{type}{1: LAI, 2:RL}
}
\value{
Default LAI and RL
}
\description{
Generate the default LAI and Roughness Length for UMD classification system
\code{tsLaiRf}
}
\examples{
tsLaiRf(type=1)
tsLaiRf(type=2)
}
