% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{VaporPressure_Sat}
\alias{VaporPressure_Sat}
\title{Saturated vapor pressure, Eq 4.2.2 in David R Maidment, Handbook of Hydrology
\code{VaporPressure_Sat}}
\usage{
VaporPressure_Sat(Temp)
}
\arguments{
\item{Temp}{Air temperature, [C]}
}
\value{
Saturated vapor pressure [kPa]
}
\description{
Saturated vapor pressure, Eq 4.2.2 in David R Maidment, Handbook of Hydrology
\code{VaporPressure_Sat}
}
