% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{voronoipolygons}
\alias{voronoipolygons}
\title{Generate Thiesson Polygons from a point data
\code{voronoipolygons}}
\usage{
voronoipolygons(x = NULL, pts = NULL, rw = NULL, crs = NULL)
}
\arguments{
\item{x}{ShapePoints* in PCS}

\item{pts}{Coordinates (x,y) of the points}

\item{rw}{The coordinates of the corners of the rectangular window enclosing the triangulation, in the order (xmin, xmax, ymin, ymax).  (default = NULL)}

\item{crs}{projection parameters (default = NULL)}
}
\value{
ShapePolygon*
}
\description{
Generate Thiesson Polygons from a point data
\code{voronoipolygons}
}
\examples{
library(rSHUD)
n = 10
xx = rnorm(n)
yy = rnorm(n)
x = sf::st_as_sf(data.frame(x = xx, y = yy), coords = c("x", "y"))
vx = voronoipolygons(x)
plot(vx, axes = TRUE)
plot(x, add = TRUE, col = 2)
#' ====END====

x = 1:5
y = 1:5
xy = expand.grid(x, y)
vx = voronoipolygons(pts = xy)
plot(vx, axes = TRUE)
points(xy)
#' ====END====

library(sf)
n = 10
xx = rnorm(n)
yy = rnorm(n)
x = sf::st_as_sf(data.frame(x = xx, y = yy), coords = c("x", "y"))
y = sf::st_as_sf(data.frame(x = xx + 2, y = yy + 2), coords = c("x", "y"))
e1 = sf::st_bbox(y)
e2 = sf::st_bbox(x)
rw = c(min(e1[1], e2[1]),
       max(e1[2], e2[2]),
       min(e1[3], e2[3]),
       max(e1[4], e2[4]) ) + c(-1, 1, -1, 1)
vx = voronoipolygons(x = x, rw = rw) 
plot(vx); plot(add = TRUE, x, col = 2); plot(add = TRUE, y, col = 3)
}
