% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelClasses.R
\docType{data}
\name{waerma}
\alias{waerma}
\title{This is Waerma Watershed example data bundled with the package}
\description{
This is Waerma Watershed example data bundled with the package
}
\keyword{data}
