% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_delineation.R
\name{watershedDelineation}
\alias{watershedDelineation}
\title{Do Watershed Delineation
\code{watershedDelineation}}
\usage{
watershedDelineation(
  fn.wbd,
  fnr.dem,
  fsp.outlets = NULL,
  dirout = tempdir(),
  dirtemp = tempdir(),
  FlowAccCell.min = NULL,
  plot = FALSE
)
}
\arguments{
\item{fn.wbd}{file of watershed SpatialPolygons}

\item{fnr.dem}{file of DEM in raster}

\item{fsp.outlets}{file of watershed outlet}

\item{dirout}{dir for out data(dem_filled.tif, dem_stm.shp, dem_wbd.shp, dem_outlets.shp) saved. Default = tempdir()}

\item{dirtemp}{dir for temporary data save. Default = tempdir()}

\item{FlowAccCell.min}{Minimum number of cells for river generation. Default  = Number of DEM / 100}
}
\value{
list of path to dem-filled, wbd, riv, outlet
}
\description{
Do Watershed Delineation
\code{watershedDelineation}
}
