% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBalance.R
\name{wb.all}
\alias{wb.all}
\title{Calculate the water balance
\code{wb.all}}
\usage{
wb.all(
  xl = loaddata(varname = c(paste0("elev", c("prcp", "eta", "etp")), "rivqdown",
    paste0("eley", c("surf", "unsat", "gw")))),
  ic = readic(),
  fun = xts::apply.monthly,
  plot = TRUE
)
}
\arguments{
\item{xl}{List of data. Five variables are included: prcp (eleqprcp), etic (eleqetic), ettr (eleqettr), etev (eleqetev) and discharge (rivqflx)}

\item{ic}{Initial Condition.}

\item{fun}{function to process the time-series data. Default = apply.daily.}

\item{plot}{Whether plot the result}
}
\value{
A matrix, contains the colums of water balance factors
}
\description{
Calculate the water balance
\code{wb.all}
}
