% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBalance.R
\name{wb.DS}
\alias{wb.DS}
\title{Calculate the Change of Storage.
\code{wb.DS}}
\usage{
wb.DS(xl = BasicPlot(varname = c(paste0("eley", c("surf", "unsat",
  "gw")), paste0("rivy", "stage")), plot = FALSE, return = TRUE),
  ic = readic())
}
\arguments{
\item{xl}{List of data. Five variables are included: surface (eleysurf), unsat zone (eleyunsat), groundwater (eleygw) and river stage (rivystage)}

\item{ic}{Initial Condition.}
}
\value{
A list. list(Ele, Riv)
}
\description{
Calculate the Change of Storage.
\code{wb.DS}
}
