% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBalance.R
\name{wb.ele}
\alias{wb.ele}
\title{Calculate the water balance
\code{wb.ele}}
\usage{
wb.ele(
  xl = loaddata(varname = c(paste0("elev", c("prcp", "etic", "ettr", "etev")),
    paste0("eleq", c("surf", "sub")), paste0("eley", c("surf", "unsat", "gw")))),
  fun = NULL,
  period = "years",
  plot = TRUE
)
}
\arguments{
\item{xl}{List of data. Five variables are included:}

\item{fun}{function to process the time-series data. Default = NULL}

\item{period}{Period of the waterbalance}

\item{plot}{Whether plot the result}
}
\value{
A matrix, contains the colums of water balance factors
}
\description{
Calculate the water balance
\code{wb.ele}
}
