% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterBalance.R
\name{wb.riv}
\alias{wb.riv}
\title{Calculate the water balance
\code{wb.riv}}
\usage{
wb.riv(
  xl = loaddata(varname = paste0("rivq", c("sub", "surf", "down"))),
  fun = NULL,
  plot = TRUE
)
}
\arguments{
\item{xl}{List of data. Five variables are included: prcp (eleqprcp),and discharge ()}

\item{fun}{function to process the time-series data. Default = NULL}

\item{plot}{Whether plot the result}
}
\value{
A matrix, contains the colums of water balance factors
}
\description{
Calculate the water balance
\code{wb.riv}
}
