% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{which_outliers}
\alias{which_outliers}
\title{Find the outliers in a dataset
\code{which_outliers}}
\usage{
which_outliers(x, na.rm = TRUE, probs = c(0.5, 0.95), ...)
}
\arguments{
\item{x}{dataset}

\item{na.rm}{Whether ignore the na value.}

\item{probs}{Range of non-outlier range}

\item{...}{More options in quatile();}
}
\value{
Index of the outliers
}
\description{
Find the outliers in a dataset
\code{which_outliers}
}
\examples{
set.seed(1)
x <- rnorm(100)
x <- c(-10, x, 10)
id <- which_outliers(x, probs=c(0.05,0.95))
y = x
y[id]=NA
par(mfrow = c(1, 2))
boxplot(x, ylim=range(x))
boxplot(y, ylim=range(x))
par(mfrow=c(1,1))
}
