% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{WindSpeed2m}
\alias{WindSpeed2m}
\title{Calculate windspeed at 2 meters.
\code{WindSpeed2m}}
\usage{
WindSpeed2m(wind, z)
}
\arguments{
\item{wind}{Windspeed [m s-1]}

\item{z}{Elevation that the windspeed was measured [m]}
}
\value{
windspeed at 2 meters
}
\description{
Calculate windspeed at 2 meters.
\code{WindSpeed2m}
}
