% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calibration_CMAES.R
\name{write.cmaes}
\alias{write.cmaes}
\title{write the configuration for CMA-ES method
\code{write.cmaes}}
\usage{
write.cmaes(x = NULL, file, backup = TRUE)
}
\arguments{
\item{x}{Configuration of CMA-ES calibration method.}

\item{file}{full path of file}

\item{backup}{TRUE/FALSE}
}
\value{
Parameters for CMA-ES configuration
}
\description{
write the configuration for CMA-ES method
\code{write.cmaes}
}
