% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{write.df}
\alias{write.df}
\title{Deprecated wrapper for \code{write.df()}}
\usage{
write.df(x, file, append = FALSE, quiet = FALSE, header = NULL)
}
\description{
Legacy compatibility wrapper for \code{\link{write_df}}.
}
