% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{write.xts}
\alias{write.xts}
\title{Write xts data out into file
\code{write.xts}}
\usage{
write.xts(x, file, append = F)
}
\arguments{
\item{x}{xts data}

\item{file}{file name}

\item{append}{whether append}
}
\description{
Write xts data out into file
\code{write.xts}
}
