% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{write_config}
\alias{write_config}
\title{Write SHUD configuration file}
\usage{
write_config(x, file)
}
\arguments{
\item{x}{SHUD model configuration parameters or calibration data}

\item{file}{Character. Full path to the output configuration file}
}
\description{
Writes SHUD model configuration parameters to a file (.para, .calib, etc.).
}
\examples{
\dontrun{
write_config(config_data, "model.para")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
