% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{write_forc}
\alias{write_forc}
\title{Write SHUD forcing file}
\usage{
write_forc(x, file, path = "", startdate = "20000101")
}
\arguments{
\item{x}{Data frame of forcing sites information with columns:
(id, lon, lat, x, y, z, filename)}

\item{file}{Character. Full path to the output .forc file}

\item{path}{Character. Common path of the forcing files (default: "")}

\item{startdate}{Character. Start date in format YYYYMMDD (default: "20000101")}
}
\description{
Writes SHUD forcing site information to a forcing file (.forc).
}
\examples{
\dontrun{
write_forc(forc_sites, "model.forc", path = "forcing/", startdate = "20100101")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
