% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{write_mesh}
\alias{write_mesh}
\title{Write SHUD mesh file}
\usage{
write_mesh(pm, file)
}
\arguments{
\item{pm}{SHUD.MESH object containing mesh and point data}

\item{file}{Character. Full path to the output .mesh file}
}
\description{
Writes a SHUD.MESH object to a mesh file (.mesh).
}
\examples{
\dontrun{
write_mesh(mesh_object, "model.mesh")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_river}()}
}
\concept{shud-io}
