% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{write_river}
\alias{write_river}
\title{Write SHUD river file}
\usage{
write_river(pr, file)
}
\arguments{
\item{pr}{SHUD.RIVER object containing river network data}

\item{file}{Character. Full path to the output .riv file}
}
\description{
Writes a SHUD.RIVER object to a river file (.riv).
}
\examples{
\dontrun{
write_river(river_object, "model.riv")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_forc_fn}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river}()},
\code{\link{read_river_sp}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()}
}
\concept{shud-io}
