% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_timeseries.R
\name{write_xts}
\alias{write_xts}
\alias{write.xts}
\title{Write xts data to file (simple format)}
\usage{
write_xts(x, file, append = FALSE)

write.xts(x, file, append = FALSE)
}
\arguments{
\item{x}{xts object containing time series data}

\item{file}{Character. Full path to output file}

\item{append}{Logical. Whether to append to existing file (default: FALSE)}
}
\description{
Writes xts time series data to file in simple format with TIME column.
This is a simpler alternative to write_tsd for basic time series output.
}
\section{Deprecated}{

\code{write.xts()} is deprecated. Use \code{write_xts()} instead.
}

\examples{
\dontrun{
# Create sample time series
dates <- seq(as.Date("2020-01-01"), as.Date("2020-12-31"), by = "day")
ts_data <- xts::xts(runif(length(dates)), order.by = dates)

# Write to simple format
write_xts(ts_data, "output.txt")
}
}
\seealso{
Other timeseries-io: 
\code{\link{read_lai}()},
\code{\link{read_tsd}()},
\code{\link{ts_to_daily}()},
\code{\link{ts_to_df}()},
\code{\link{write_tsd}()}
}
\concept{timeseries-io}
