% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{xy2ID}
\alias{xy2ID}
\title{Convert the \code{xy} (X,Y) into Index in \code{coords}
\code{xy2ID}}
\usage{
xy2ID(xy, coord)
}
\arguments{
\item{xy}{coordinate (x,y), subset of \code{coords}}

\item{coord}{coordinate set}
}
\value{
Index in coodinate set
}
\description{
Convert the \code{xy} (X,Y) into Index in \code{coords}
\code{xy2ID}
}
\examples{
coord = matrix(rnorm(10), 5,2)
xy = coord[c(5,1,3),]
xy2ID(xy, coord)
}
