% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetCDF.R
\name{xyz2Raster}
\alias{xyz2Raster}
\title{Read the time tage from (x, y, matrix).}
\usage{
xyz2Raster(x, y = NULL, arr = NULL, flip = TRUE, plot = TRUE)
}
\arguments{
\item{x}{list(x, y, arr) or x coordinates}

\item{y}{y coordinates}

\item{arr}{array. dim = (nx, ny)}

\item{flip}{Whether flip the matrix.}

\item{plot}{Whether plot the raster.}
}
\value{
a Raster/RasterStack
}
\description{
Read the time tage from (x, y, matrix).
}
