% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{AerodynamicResistance}
\alias{AerodynamicResistance}
\title{Aerodynamic Resistance. Eq 4.2.25 in David R Maidment, Handbook of Hydrology
\code{AerodynamicResistance}}
\usage{
AerodynamicResistance(Uz, hc, Z_u, Z_e, VON_KARMAN = 0.4)
}
\arguments{
\item{Uz}{Wind speed at height z [m s-1]}

\item{hc}{Height of crop [m]}

\item{Z_u}{height of wind measurements [m]}

\item{Z_e}{height of humidity measurements [m]}

\item{VON_KARMAN}{von Karman's constant. Default = 0.41 [-]}
}
\value{
Aerodynamic resistance [s m-1]
}
\description{
Aerodynamic Resistance. Eq 4.2.25 in David R Maidment, Handbook of Hydrology
\code{AerodynamicResistance}
}
