% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{AirDensity}
\alias{AirDensity}
\title{Air Density, Eq 4.2.4 in David R Maidment, Handbook of Hydrology
\code{shud.ic}}
\usage{
AirDensity(P, Temp)
}
\arguments{
\item{P}{Air pressure [kP]}

\item{Temp}{Air temperature [C]}
}
\value{
Density of Air. [kg m-3]
}
\description{
Air Density, Eq 4.2.4 in David R Maidment, Handbook of Hydrology
\code{shud.ic}
}
