% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_main.R
\name{autoBuildModel}
\alias{autoBuildModel}
\title{Deprecated: autoBuildModel}
\usage{
autoBuildModel(...)
}
\arguments{
\item{...}{Arguments (not supported - will show error with migration guidance)}
}
\description{
This function is deprecated. Use \code{\link{auto_build_model}} instead.
The new function uses modern spatial libraries (terra/sf) exclusively.
}
\keyword{internal}
