% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoBuildModel.R
\name{autoBuildModel}
\alias{autoBuildModel}
\title{Automatically run SHUDgis based on the input data and parameters.
\code{autoBuildModel}}
\usage{
autoBuildModel(
  indata,
  forcfiles,
  prjname = "sac",
  outdir = getwd(),
  a.max = 1e+06 * 0.2,
  q.min = 33,
  tol.riv = 0,
  tol.wb = 200,
  tol.len = 0,
  AqDepth = 10,
  years = 2000:2010,
  clean = TRUE,
  cfg.para = shud.para(nday = 365 * length(years) + round(length(years)/4)),
  cfg.calib = shud.calib(),
  mf = MeltFactor(years = years),
  rm.outlier = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{indata}{Input data, list of data.}

\item{forcfiles}{Filenames of .csv forcing data}

\item{prjname}{Projectname}

\item{outdir}{Output directory}

\item{a.max}{Maximum area of triangles. Unit=m2}

\item{q.min}{Minimum of angles for each triangle}

\item{tol.riv}{Tolerance to simplify river lines, in meter}

\item{tol.wb}{Tolerance to simplify watershed boundary, in meter}

\item{tol.len}{Tolerance to simplyfy river by longituanal length, meter.}

\item{AqDepth}{Depth of aquifer bottom, which is surface elemvation minus bedrock elevation. meter.}

\item{years}{Years to generate the LAI, Meltfactor, Roughness Length, which must be same as the period of forcing data.}

\item{clean}{Whether clean the existing model files in output directory.}

\item{cfg.para}{model configuration, parameter}

\item{cfg.calib}{model calibration}

\item{mf}{Meltfactor}

\item{rm.outlier}{Whether to remove the outlier in soil/geol;}

\item{quiet}{Whether to ask confirmation when multiple outlets exist.}
}
\value{
\code{SHUD.mesh}
}
\description{
Automatically run SHUDgis based on the input data and parameters.
\code{autoBuildModel}
}
\examples{
data(sac)
indata = sac
sp.forc =indata[['forc']]
forc.fns = paste0(sp.forc@data[, 'NLDAS_ID'], '.csv')
forc.fns
}
