% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readout.R
\name{BasicPlot}
\alias{BasicPlot}
\title{Read multiply SHUD model output files and do time-series plot}
\usage{
BasicPlot(
  varname = c(paste0("eley", c("surf", "unsat", "gw", "snow")), paste0("elev", c("prcp",
    "infil", "rech")), paste0("elev", c("etp", "eta", "etev", "ettr", "etic")),
    paste0("rivq", c("down", "sub", "surf")), paste0("rivy", "stage")),
  xl = NULL,
  sp.riv = NULL,
  rdsfile = NULL,
  plot = TRUE,
  imap = FALSE,
  w.focal = matrix(1/9, 3, 3),
  path = file.path(get("anapath", envir = .shud), "BasicPlot")
)
}
\arguments{
\item{varname}{vector of output keywords}

\item{xl}{list of data returned from loaddata()}

\item{sp.riv}{River SpatialLine*}

\item{rdsfile}{Save RDS file}

\item{plot}{Whether do the time-series plot}

\item{imap}{Whether do the raster plot for Element data. Only works for element data}

\item{w.focal}{forcal matrix}

\item{return}{Whether return the data. Some the results are too huge to load in memoery at once.}

\item{iRDS}{Whether save RDS file.}
}
\value{
A list of TimeSeries data.
}
\description{
Read multiply SHUD model output files and do time-series plot
}
\keyword{output.}
\keyword{read}
