% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{CanopyResistance}
\alias{CanopyResistance}
\title{Canopy resistance
\code{CanopyResistance}}
\usage{
CanopyResistance(rmin, lai)
}
\arguments{
\item{rmin}{Minimum Resistance [s m-1]}
}
\value{
lai Leaf Area Index (LAI) [m2 m-2]
}
\description{
Canopy resistance
\code{CanopyResistance}
}
