% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{count}
\alias{count}
\title{Count the frequency of numbers
\code{count}}
\usage{
count(x, n = NULL)
}
\arguments{
\item{x}{target value}

\item{n}{specific frequency}
}
\value{
count of specific frequency
}
\description{
Count the frequency of numbers
\code{count}
}
\examples{
x=c(round(rnorm(100, 2)+1), 0,0)
count(x)
count(x, sum(x==0))
}
