% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_Projection.R
\name{crs.Albers}
\alias{crs.Albers}
\title{Build an Albers Equal Area projection based on the spatial data or extent.
\code{crs.Albers}}
\usage{
crs.Albers(spx, ext = NULL)
}
\arguments{
\item{spx}{Spatial Data}

\item{ext}{extent of the data.}
}
\value{
CRS in Albers Equal Area projection.
}
\description{
Build an Albers Equal Area projection based on the spatial data or extent.
\code{crs.Albers}
}
\examples{
crs.Albers(ext=c(90, 100, 35, 40))
}
