% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_Projection.R
\name{crs.Lambert}
\alias{crs.Lambert}
\title{Build a Lambert Equal Area projection based on the spatial data or extent.
\code{crs.Lambert}}
\usage{
crs.Lambert(spx, ext = NULL)
}
\arguments{
\item{spx}{Spatial Data}

\item{ext}{extent of the data.}
}
\value{
CRS in Lambert
}
\description{
Build a Lambert Equal Area projection based on the spatial data or extent.
\code{crs.Lambert}
}
\examples{
crs.Lambert(ext=c(90, 100, 35, 40))
}
