% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_Projection.R
\name{crs.long2utm}
\alias{crs.long2utm}
\title{Get UTM projection parameters in CRS, from longtitude and latitude.
\code{crs.long2utm}}
\usage{
crs.long2utm(lon, lat = 30)
}
\arguments{
\item{lon}{Longitude in degree}

\item{lat}{Latitude in degree, default = 30, North.}
}
\value{
UTM projection parameters in CRS
}
\description{
Get UTM projection parameters in CRS, from longtitude and latitude.
\code{crs.long2utm}
}
\examples{
long = seq(-180, 179, 6)+1
x =crs.long2utm(long)
}
