% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{extractCoords}
\alias{extractCoords}
\title{Deprecated wrapper for \code{extractCoords()}}
\usage{
extractCoords(...)
}
\description{
Legacy compatibility wrapper for \code{\link{get_coords}}.
}
