% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{extractRaster}
\alias{extractRaster}
\title{Extract values on Raster map along a normalized line from 0 to 1.
\code{extractRaster}}
\usage{
extractRaster(r, xy = NULL, ext = NULL, plot = TRUE)
}
\arguments{
\item{r}{Raster}

\item{xy}{Coordinates of the line, dim = (Npoints, 2); x and y must be in
the range 0 to 1.}

\item{ext}{extension of value xy.}

\item{plot}{Whether plot result.}
}
\description{
Extract values on Raster map along a normalized line from 0 to 1.
\code{extractRaster}
}
\examples{
r <- terra::rast(ncols = 36, nrows = 18)
terra::values(r) <- seq_len(terra::ncell(r))
extractRaster(r)
}
