% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{extractRaster}
\alias{extractRaster}
\title{Extract values on Raster map. The line is a straight line between (0,1). 
\code{extractRaster}}
\usage{
extractRaster(r, xy = NULL, ext = raster::extent(r), plot = T)
}
\arguments{
\item{r}{Raster}

\item{xy}{coordinates of the line, dim=(Npoints, 2); x and y must be in [0, 1]}

\item{ext}{extension of value xy.}

\item{plot}{Whether plot result.}
}
\description{
Extract values on Raster map. The line is a straight line between (0,1). 
\code{extractRaster}
}
\examples{
library(raster)
}
