% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{ForcingCoverage}
\alias{ForcingCoverage}
\title{Generate the coverage map for forcing sites.
\code{ForcingCoverage}}
\usage{
ForcingCoverage(
  sp.meteoSite = NULL,
  filenames = paste0(sp.meteoSite@data$ID, ".csv"),
  pcs,
  gcs = sp::CRS("+init=epsg:4326"),
  dem,
  wbd,
  enlarge = 10000
)
}
\arguments{
\item{sp.meteoSite}{ShapePoints* in PCS}

\item{pcs}{Projected Coordinate System}

\item{gcs}{Geographic Coordinate System}

\item{dem}{DEM raster}

\item{wbd}{watershed boundary}

\item{enlarge}{enlarge factor for the boundary.}
}
\value{
ShapePolygon*
}
\description{
Generate the coverage map for forcing sites.
\code{ForcingCoverage}
}
