% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{FromToNode}
\alias{FromToNode}
\title{return the FROM and TO nodes index of the SpatialLines
\code{FromToNode}}
\usage{
FromToNode(sp, coord = extractCoords(sp, unique = TRUE))
}
\arguments{
\item{sp}{SpatialLines*}

\item{coord}{Coordinate of vertex in \code{sp}}
}
\value{
FROM and TO nodes index of the SpatialLines
}
\description{
return the FROM and TO nodes index of the SpatialLines
\code{FromToNode}
}
