% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getRiverNodes}
\alias{getRiverNodes}
\title{Get the From/To nodes of the river.
\code{getRiverNodes}}
\usage{
getRiverNodes(spr = read_river_sp())
}
\arguments{
\item{spr}{River line geometry. Defaults to the compatibility
\code{\link{read_river_sp}()} reader; \code{sf} LINESTRING input is also
accepted.}
}
\value{
a list, c(points, FT_ID)
}
\description{
Get the From/To nodes of the river.
\code{getRiverNodes}
}
