% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Get.R
\name{getRiverNodes}
\alias{getRiverNodes}
\title{Get the From/To nodes of the river.
\code{getRiverNodes}}
\usage{
getRiverNodes(spr = readriv.sp())
}
\arguments{
\item{spr}{SpatialLine* of river streams.}
}
\value{
a list, c(points, FT_ID)
}
\description{
Get the From/To nodes of the river.
\code{getRiverNodes}
}
