% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{grid.subset}
\alias{grid.subset}
\title{Find the subset of a extent in a grid.
\code{grid.subset}}
\usage{
grid.subset(
  ext,
  res,
  ext.sub,
  dx = matrix(res, 2, 1)[1],
  dy = matrix(res, 2, 1)[2],
  x = seq(ext[1] + dx/2, ext[2] - dx/2, by = dx),
  y = seq(ext[3] + dy/2, ext[4] - dy/2, by = dy)
)
}
\arguments{
\item{ext}{txtent of the grid}

\item{res}{resolution of the grid}

\item{ext.sub}{the extent of subset}

\item{x}{x coordinates of the grids}

\item{y}{y coordinates of the grids}
}
\value{
list, list(xid, yid, x, y)
}
\description{
Find the subset of a extent in a grid.
\code{grid.subset}
}
