% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{highlight_id}
\alias{highlight_id}
\title{Highlight elements with id.
\code{highlight_id}}
\usage{
highlight_id(EleID = NULL, RivID = NULL)
}
\arguments{
\item{EleID}{Index of element}

\item{RivID}{Index of river reaches}
}
\description{
Highlight elements with id.
\code{highlight_id}
}
