% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hydrograph}
\alias{hydrograph}
\title{Plot hydrograph
\code{hydrograph}}
\usage{
hydrograph(x, legend.position = "bottom", unit = rep("", ncol(x)),
  col = c(3, 4), heights = c(3, 7), bg = TRUE, ylabs = NULL)
}
\arguments{
\item{x}{time-seres matrix. The first column will be plot on top subfigure.}

\item{legend.position}{Location to put the legend for discharge.}

\item{unit}{Unit of the variables.}

\item{col}{colors of each variable.}

\item{heights}{Heights of top (rainfall) figure and bottom (discharge) figure}

\item{bg}{Whether plot the background rectangles.}

\item{ylabs}{ylab for top and bottom plot.}
}
\description{
Plot hydrograph
\code{hydrograph}
}
