% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{map2d}
\alias{map2d}
\title{Raster plot the mesh data
\code{map2d}}
\usage{
map2d(x = getElevation(), sp.riv = NULL)
}
\arguments{
\item{x}{Vector or Matrix of mesh data. Length of nrow must be equal the number of cells}

\item{sp.riv}{The SpatialLines of river network. sp.riv=NULL means plot raster only.}
}
\value{
Raster of the x (vector) or the last row of x (matrix)
}
\description{
Raster plot the mesh data
\code{map2d}
}
