% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_PTF.R
\name{MeltFactor}
\alias{MeltFactor}
\title{Generate the default melt factor
\code{MeltFactor}}
\usage{
MeltFactor(
  years,
  mf = c(0.001308019, 0.001633298, 0.002131198, 0.002632776, 0.003031171, 0.003197325,
    0.003095839, 0.00274524, 0.002260213, 0.001759481, 0.001373646, 0.001202083)
)
}
\arguments{
\item{years}{Years.}

\item{mf}{Default monthly melting factors.}
}
\value{
Multiple years monthly melting factor.
}
\description{
Generate the default melt factor
\code{MeltFactor}
}
\examples{
years = 2000:2001
x = MeltFactor(years=years)
plot(x)
}
