% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelInfo.R
\name{MeshAtt}
\alias{MeshAtt}
\title{Summary of the attributions on Mesh
\code{MeshAtt}}
\usage{
MeshAtt(pm = readmesh(), att = readatt(), soil = readsoil(),
  geol = readgeol(), lc = readlc())
}
\arguments{
\item{pm}{SHUD.MESH, .sp.mesh}

\item{att}{.sp.att of input}

\item{soil}{.para.soil of input}

\item{geol}{.para.geol of input}

\item{lc}{.para.lc of input}
}
\value{
Attributes of each element.
}
\description{
Summary of the attributions on Mesh
\code{MeshAtt}
}
