% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{MeshData2Raster}
\alias{MeshData2Raster}
\title{SpatialData to Raster
\code{MeshData2Raster}}
\usage{
MeshData2Raster(
  x = getElevation(),
  rmask = shud.mask(proj = proj),
  pm = readmesh(),
  proj = NULL,
  stack = FALSE,
  method = "ide",
  plot = FALSE
)
}
\arguments{
\item{x}{vector or matrix, length/nrow is number of cells.}

\item{rmask}{mask of the mesh file}

\item{pm}{shud mesh}

\item{proj}{Projejction parameter}

\item{stack}{Whether export the stack, only when the x is a matrix, i.e. (Ntime x Ncell).}

\item{plot}{Whether plot the result.}
}
\value{
Raster map
}
\description{
SpatialData to Raster
\code{MeshData2Raster}
}
