% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{NodeIDList}
\alias{NodeIDList}
\title{return the Index of nodes in SpatialData
\code{NodeIDList}}
\usage{
NodeIDList(sp, coord = extractCoords(sp, unique = TRUE))
}
\arguments{
\item{sp}{SpatialLines*}

\item{coord}{Coordinate of vertex in \code{sp}}
}
\value{
list of Index of each SpatialData, line or polygon
}
\description{
return the Index of nodes in SpatialData
\code{NodeIDList}
}
