% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{PET_PM}
\alias{PET_PM}
\title{Potential Evapotranspiration with Pennmann-Monteith equation.
\code{PET_PM}}
\usage{
PET_PM(
  Wind,
  Temp,
  RH,
  RadNet,
  Press,
  WindHeight = 10,
  Veg_height = 0.12,
  albedo = 0.23,
  Res_surf = 70,
  Elevation = NULL
)
}
\arguments{
\item{Wind}{Windspeed [m s-1]}

\item{Temp}{Air temperature, [C]}

\item{RH}{Relative Humidity [-] 0~1.}

\item{RadNet}{Net Solar Radiation [w m-2]}

\item{Press}{Pressure [kPa]}

\item{WindHeight}{Height that wind was measured [m]}

\item{Veg_height}{Height of vegetation [m]}

\item{albedo}{Albedo [-]}

\item{Res_surf}{Aerodynamic surface resistance [s m-1]}

\item{Elevation}{Elevation [m]}
}
\value{
PET at [mm m-2 s-1] or [kg m-2 s-1]
}
\description{
Potential Evapotranspiration with Pennmann-Monteith equation.
\code{PET_PM}
}
