% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plot_sp}
\alias{plot_sp}
\title{Plot SpatialPolygons maps
\code{plot.sp}}
\usage{
plot_sp(x, zcol, col.fun = topo.colors, ncol = 20, ...)
}
\arguments{
\item{x}{Spatial*}

\item{zcol}{The column to plot}

\item{col.fun}{Functions of coloring.}

\item{ncol}{Number of colors in color function.}

\item{...}{more options in raster::plot()}
}
\description{
Plot SpatialPolygons maps
\code{plot.sp}
}
