% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot time-series hydrograph}
\usage{
plot_timeseries(x, ...)
}
\arguments{
\item{x}{xts or zoo time-series matrix.}

\item{...}{Additional arguments passed to \code{\link{plot_hydrograph}}.}
}
\value{
A combined plot object from \code{\link{plot_hydrograph}}.
}
\description{
Compatibility alias for \code{\link{plot_hydrograph}} retained for code that
used the transitional \code{plot_timeseries()} name.
}
\examples{
\dontrun{
library(xts)
dates <- as.POSIXct(as.Date("2020-01-01") + 1:10)
x <- xts(cbind(precip = 1:10, discharge = 11:20), order.by = dates)
plot_timeseries(x)
}
}
