% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{plot_tsd}
\alias{plot_tsd}
\title{Deprecated wrapper for \code{plot_tsd()}}
\usage{
plot_tsd(x, ...)
}
\arguments{
\item{x}{Time-series object to plot.}

\item{...}{Additional arguments passed to \code{\link{plot_timeseries}}.}
}
\value{
A plot object from \code{\link{plot_timeseries}}.
}
\description{
Legacy compatibility wrapper for \code{\link{plot_timeseries}}.
}
