% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{plot_tsd}
\alias{plot_tsd}
\title{Deprecated wrapper for \code{plot_tsd()}}
\usage{
plot_tsd(x, ...)
}
\arguments{
\item{x}{Time-series object to plot.}

\item{...}{Additional arguments passed to the selected plotting function.}
}
\value{
A plot object from \code{\link{plot_timeseries}} for hydrograph
inputs, otherwise \code{x} invisibly.
}
\description{
Legacy compatibility wrapper for \code{\link{plot_timeseries}}. Hydrograph
inputs are forwarded to \code{\link{plot_timeseries}}; other base-plottable
inputs use \code{\link[graphics]{plot}} to preserve legacy behavior.
}
