% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tsd}
\alias{plot_tsd}
\title{Plot xts data..
\code{plot_tsd}}
\usage{
plot_tsd(x, time.col = "year")
}
\arguments{
\item{x}{xts data}

\item{time.col}{Time interval of banded background. Default is 'year'}
}
\description{
Plot xts data..
\code{plot_tsd}
}
\examples{
library(xts)
nday = 1000
xd=as.POSIXct(as.Date('2000-01-01')+ 1:nday )
x=as.xts(sin(1:1000 / 100), order.by=xd)
plot_tsd(x)
}
