% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{png.control}
\alias{png.control}
\title{Prepare png figure out.
\code{png.control}}
\usage{
png.control(fn = "figure.png", file = file.path(path, fn),
  path = get("anapath", envir = .shud), ratio = 9/11, ht = 11,
  wd = ht * ratio, units = "in", res = 200)
}
\arguments{
\item{fn}{Filename}

\item{file}{Full path of file}

\item{path}{The path only. Default = anapath in .shud environment.}

\item{ratio}{Ratio of width to height.}

\item{ht}{Height.}

\item{wd}{Width.}

\item{units}{Units of width and height.}

\item{res}{Resolution along height and width.}
}
\description{
Prepare png figure out.
\code{png.control}
}
