% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{PointInDistance}
\alias{PointInDistance}
\title{Find the points in distance less than tol.
\code{PointInDistance}}
\usage{
PointInDistance(pt, tol)
}
\arguments{
\item{pt}{2-column coordinates (x,y).}

\item{tol}{Tolerance}
}
\value{
Index of points that within tol
}
\description{
Find the points in distance less than tol.
\code{PointInDistance}
}
