% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{PsychrometricConstant}
\alias{PsychrometricConstant}
\title{Psychrometric Constant, Eq 4.2.1 in David R Maidment, Handbook of Hydrology
\code{PsychrometricConstant}}
\usage{
PsychrometricConstant(Pressure, lambda)
}
\arguments{
\item{Pressure}{Air pressure [kPa]}

\item{lambda}{Latent heat of vaporized water. [MJ/kg]}
}
\value{
γ psychrometric constant [kPa °C-1]
}
\description{
Psychrometric Constant, Eq 4.2.1 in David R Maidment, Handbook of Hydrology
\code{PsychrometricConstant}
}
