% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_shud.R
\name{read_forc_fn}
\alias{read_forc_fn}
\alias{readforc.fn}
\title{Read SHUD forcing file list}
\usage{
read_forc_fn(file = shud.filein()["md.forc"])

readforc.fn(file = shud.filein()["md.forc"])
}
\arguments{
\item{file}{Character. Full path to the .forc file}
}
\value{
List containing start time and forcing sites data frame
}
\description{
Reads a SHUD forcing file (.forc) and returns forcing site information.
}
\section{Deprecated}{

\code{readforc.fn()} is deprecated. Use \code{read_forc_fn()} instead.
}

\examples{
\dontrun{
forc <- read_forc_fn("model.forc")
}
}
\seealso{
Other shud-io: 
\code{\link{read_att}()},
\code{\link{read_calib}()},
\code{\link{read_config}()},
\code{\link{read_df}()},
\code{\link{read_forc_csv}()},
\code{\link{read_geol}()},
\code{\link{read_ic}()},
\code{\link{read_lc}()},
\code{\link{read_mesh}()},
\code{\link{read_para}()},
\code{\link{read_river_sp}()},
\code{\link{read_river}()},
\code{\link{read_rivseg}()},
\code{\link{read_soil}()},
\code{\link{write_config}()},
\code{\link{write_df}()},
\code{\link{write_forc}()},
\code{\link{write_ic}()},
\code{\link{write_mesh}()},
\code{\link{write_river}()}
}
\concept{shud-io}
