% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_timeseries.R
\name{read_lai}
\alias{read_lai}
\alias{readlai}
\title{Read LAI time series file}
\usage{
read_lai(file = shud.filein()["md.lai"])

readlai(file = shud.filein()["md.lai"])
}
\arguments{
\item{file}{Character. Full path to LAI time series file}
}
\value{
List of time-series data. One-block files are named \code{LAI}; two-block
files are named \code{LAI} and \code{RL} (Roughness Length). Additional blocks, if
present, are named \code{Block3}, \code{Block4}, etc.
}
\description{
Reads a LAI (Leaf Area Index) time series file (.tsd.lai or legacy
.ts.lai) and returns properly named time-series components. Real SHUD input
files may contain only one block; those files are returned as \code{LAI}. Two-block
compatibility files are returned as \code{LAI} and \code{RL}.
}
\section{Deprecated}{

\code{readlai()} is deprecated. Use \code{read_lai()} instead.
}

\examples{
\dontrun{
lai_data <- read_lai("model.ts.lai")
plot(lai_data$LAI)
if ("RL" \%in\% names(lai_data)) plot(lai_data$RL)
}
}
\seealso{
Other timeseries-io: 
\code{\link{read_tsd}()},
\code{\link{ts_to_daily}()},
\code{\link{ts_to_df}()},
\code{\link{write_tsd}()},
\code{\link{write_xts}()}
}
\concept{timeseries-io}
