% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readout.R
\name{readout}
\alias{readout}
\title{Read output file from SHUD model}
\usage{
readout(
  keyword,
  path = get("outpath", envir = .shud),
  ASCII = FALSE,
  ver = get("version", envir = .shud),
  file = file.path(path, paste0(get("PRJNAME", envir = .shud), ".", keyword, ".dat"))
)
}
\arguments{
\item{keyword}{keyword of the output file. File name format is: projectname.keyword.dat}

\item{path}{path of the output file}

\item{ASCII}{If TRUE, convert the binary file to ASCII format (.csv)}

\item{ver}{Version of output file. Ver = 1.0, 2.0}

\item{file}{Full path of output file.}
}
\value{
A Time-Series data. This list require support of xts packages.
}
\description{
Read output file from SHUD model
}
\keyword{Could}
\keyword{and}
\keyword{be}
\keyword{for}
\keyword{format.}
\keyword{mesh}
\keyword{output.}
\keyword{read}
\keyword{reading}
\keyword{river}
\keyword{used}
