% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{rmDuplicatedLines}
\alias{rmDuplicatedLines}
\title{Remove the duplicated lines, which means the FROM and TO points are identical.
\code{rmDuplicatedLines}}
\usage{
rmDuplicatedLines(x, ...)
}
\arguments{
\item{x}{ShapeLine*}

\item{...}{More options in duplicated()}
}
\value{
ShapeLine* without duplicated lines.
}
\description{
Remove the duplicated lines, which means the FROM and TO points are identical.
\code{rmDuplicatedLines}
}
