% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{rowMatch}
\alias{rowMatch}
\title{Identify whether every row of x is same as m;}
\usage{
rowMatch(x, m)
}
\arguments{
\item{x}{Row vector}

\item{m}{Matrix}
}
\value{
TRUE/FALSE of matching
}
\description{
Identify whether every row of x is same as m;
}
\examples{
rowMatch(c(1, 1), cbind(1:10, 1:10))
}
