% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{SharedPoints}
\alias{SharedPoints}
\title{Find shared points in line geometries
\code{SharedPoints}}
\usage{
SharedPoints(sp, plot = FALSE)
}
\arguments{
\item{sp}{\code{sf} LINESTRING object; legacy \code{SpatialLines} is accepted
for compatibility.}

\item{plot}{Whether to plot the results for debugging}
}
\value{
Topologic relationship between line components.
}
\description{
Find shared points in line geometries
\code{SharedPoints}
}
