% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelClasses.R
\docType{class}
\name{SHUD River-class}
\alias{SHUD River-class}
\alias{SHUD.RIVER}
\title{Class of SHUD.RIVER
\code{SHUD.RIVER}}
\value{
Class of SHUD.RIVER
}
\description{
Class of SHUD.RIVER
\code{SHUD.RIVER}
}
\section{Slots}{

\describe{
\item{\code{river}}{data.frame or matrix}

\item{\code{rivertype}}{data.frame or matrix}

\item{\code{point}}{data.frame or matrix}
}}

