% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeshDomain.R
\name{shud.att}
\alias{shud.att}
\title{extract Coordinates of SpatialLines or  SpatialPolygons
\code{shud.att}}
\usage{
shud.att(tri, r.soil = NULL, r.geol = NULL, r.lc = NULL,
  r.forc = NULL, r.mf = NULL, r.BC = NULL, r.SS = NULL)
}
\arguments{
\item{tri}{Triangles defination}

\item{r.soil}{raster of soil classification}

\item{r.geol}{raster of geology layer}

\item{r.lc}{raster of land cover, LAI and Roughness Length}

\item{r.forc}{raster of forcing data}

\item{r.mf}{raster of melt factor}

\item{r.BC}{raster of boundary condition}

\item{r.SS}{raster of Source/Sink}
}
\value{
data.frame of SHUD .att
}
\description{
extract Coordinates of SpatialLines or  SpatialPolygons
\code{shud.att}
}
