% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{shud.mask}
\alias{shud.mask}
\title{Generate the raster mask of Mesh domain
\code{shud.mask}}
\usage{
shud.mask(pm = readmesh(), proj = NULL, rr = get("MASK", envir =
  .shud), n = 40000, cellsize = NULL)
}
\arguments{
\item{pm}{\code{shud.mesh}}

\item{proj}{Projection parameter}

\item{rr}{Default mask in .shud environment}

\item{n}{Number of grid}

\item{cellsize}{Resolution, defaul = NULL, resolution = extent / ngrids}
}
\value{
Raster map
}
\description{
Generate the raster mask of Mesh domain
\code{shud.mask}
}
