% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeshDomain.R
\name{shud.mesh}
\alias{shud.mesh}
\title{Generate the mesh data from the triangulation
\code{shud.mesh}}
\usage{
shud.mesh(tri, dem, AqDepth = 10, r.aq = dem * 0 + AqDepth)
}
\arguments{
\item{tri}{Triangles defination}

\item{dem}{Elevation. Projection of the DEM raster must be same as watershed boundary.}

\item{AqDepth}{Aquifer depth, numeric.}

\item{r.aq}{Aquifer Thickness. Raster object}
}
\value{
Triangle mesh of model domain.
}
\description{
Generate the mesh data from the triangulation
\code{shud.mesh}
}
\examples{
library(raster)
data(sac)
wbd=sac[['wbd']]
dem=sac[['dem']]
a.max = 1e6 * 1;
q.min = 33;
tol.riv = 200
tol.wb = 200
tol.len = 500
AqDepth = 10

wbbuf = rgeos::gBuffer(wbd, width = 2000)
dem = raster::crop(dem, wbbuf)

wb.dis = rgeos::gUnionCascaded(wbd)
wb.simp = rgeos::gSimplify(wb.dis, tol=tol.wb, topologyPreserve = TRUE)


tri = shud.triangle(wb=wb.simp,q=q.min, a=a.max)
plot(tri, asp=1)

# generate SHUD .mesh
pm=shud.mesh(tri,dem=dem, AqDepth = AqDepth)
sm = sp.mesh2Shape(pm)
raster::plot(sm)
}
