% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{shud.river}
\alias{shud.river}
\title{calculate river order, downstream, slope and length
\code{shud.river}}
\usage{
shud.river(sl, dem)
}
\arguments{
\item{sl}{SpatialLines*}

\item{dem}{Raster of elevation}
}
\value{
SHUD.RIVER
}
\description{
calculate river order, downstream, slope and length
\code{shud.river}
}
