% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{shud.rivseg}
\alias{shud.rivseg}
\title{Generate the river segments table
\code{shud.rivseg}}
\usage{
shud.rivseg(sl)
}
\arguments{
\item{sl}{SpatialLinesDataFrame}
}
\value{
river segments table
}
\description{
Generate the river segments table
\code{shud.rivseg}
}
