% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglulate.R
\name{shud.triangle}
\alias{shud.triangle}
\title{Generate triangle mesh domain
\code{shud.triangle}}
\usage{
shud.triangle(wb, dem = NULL, riv = NULL, hole = NULL, pts = NULL, q = 30, ...)
}
\arguments{
\item{wb}{SpatialPolygon or SpatialLines which define the watershed boundary}

\item{dem}{Elevation data.}

\item{riv}{SpatialLines of river network, optional}

\item{pts}{Extra pts to build triangular mesh.}

\item{q}{minimum angle of triangle}

\item{...}{more options in RTriangle::triangulate()}

\item{lake}{SpatialPolygon of lake.}
}
\value{
A object with class triangulation.
}
\description{
Generate triangle mesh domain
\code{shud.triangle}
}
