% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_SimplifyPolylines.R
\name{SimplifybyLen}
\alias{SimplifybyLen}
\title{Simplify Polylines by length
\code{SimplifybyLen}}
\source{
https://stackoverflow.com/questions/38700246/how-do-i-split-divide-polyline-shapefiles-into-equally-length-smaller-segments
}
\usage{
SimplifybyLen(sp, split_length = 20, plot.results = F)
}
\arguments{
\item{sp}{a SpatialLines or SpatialPoygons object}

\item{split_length}{the length of the segments to split the lines into, in units of the SpatialLine object}

\item{plot.results}{TRUE/FALSE}
}
\value{
simplified SpatialLines
}
\description{
Simplify Polylines by length
\code{SimplifybyLen}
}
