% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{SinglePolygon}
\alias{SinglePolygon}
\title{Conver the MULTIPOLYGONS to SINGLEPOLYGONS.}
\usage{
SinglePolygon(x, id = 0)
}
\arguments{
\item{x}{the spatialpolygon*}

\item{id}{Index of the sorted (decreasing) polygons to return. default = 0;}
}
\description{
Conver the MULTIPOLYGONS to SINGLEPOLYGONS.
}
