% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{SlopeSatVaporPressure}
\alias{SlopeSatVaporPressure}
\title{Slope of saturated vapor pressure, Eq 4.2.3 in David R Maidment, Handbook of Hydrology
\code{SlopeSatVaporPressure}}
\usage{
SlopeSatVaporPressure(Temp, ES)
}
\arguments{
\item{Temp}{Tempearture [C]}

\item{ES}{}
}
\value{

}
\description{
Slope of saturated vapor pressure, Eq 4.2.3 in David R Maidment, Handbook of Hydrology
\code{SlopeSatVaporPressure}
}
