% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglulate.R
\name{sp.mesh2Shape}
\alias{sp.mesh2Shape}
\title{Generate triangle mesh domain
\code{sp.mesh2Shape}}
\usage{
sp.mesh2Shape(pm = readmesh(), dbf = NULL, crs = NULL)
}
\arguments{
\item{pm}{SHUD mesh object}

\item{dbf}{attribute table of the mesh triangles.}

\item{crs}{Projection parameters.}
}
\value{
SpatialPolygons object
}
\description{
Generate triangle mesh domain
\code{sp.mesh2Shape}
}
