% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/River.R
\name{sp.riv2shp}
\alias{sp.riv2shp}
\title{Convert the SHUD.RIVER class to SpatialLines
\code{sp.riv2shp}}
\usage{
sp.riv2shp(pr = readriv(), dbf = NULL)
}
\arguments{
\item{pr}{SHUD.RIVER class}

\item{dbf}{Attribute data for exported SpatialLines}
}
\value{
SpatialLinesDataFrame
}
\description{
Convert the SHUD.RIVER class to SpatialLines
\code{sp.riv2shp}
}
