% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{sp.RiverDown}
\alias{sp.RiverDown}
\title{determine index of downstream
\code{sp.RiverOrder}}
\usage{
sp.RiverDown(sp, coord = extractCoords(sp))
}
\arguments{
\item{sp}{SpatialLines*}

\item{coord}{coordinates of \code{sp}.}
}
\value{
Index of downstream for each segements
}
\description{
determine index of downstream
\code{sp.RiverOrder}
}
