% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{sp.RiverOrder}
\alias{sp.RiverOrder}
\title{calculate river order
\code{sp.RiverOrder}}
\usage{
sp.RiverOrder(sp, coord = extractCoords(sp))
}
\arguments{
\item{sp}{SpatialLines*}

\item{coord}{coordinates of \code{sp}.}
}
\value{
Stream Order of SpatialLines*
}
\description{
calculate river order
\code{sp.RiverOrder}
}
\examples{
library(rgdal)
data(sac)
riv=sac$riv
ord = sp.RiverOrder(riv)
print(ord)
idx = sort(unique(ord))
raster::plot(riv, col=ord)
legend('topleft', legend=idx, col=idx, lwd=1, lty=1)
}
