% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_RiverProcess.R
\name{sp.RiverSeg}
\alias{sp.RiverSeg}
\title{Cut the river by shud.mesh
\code{sp.RiverSeg}}
\usage{
sp.RiverSeg(sp.mesh, sp.riv)
}
\arguments{
\item{sp.mesh}{shud mesh}

\item{sp.riv}{river}
}
\value{
SpatialLinesDataFrame
}
\description{
Cut the river by shud.mesh
\code{sp.RiverSeg}
}
