% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS_SimplifybyLen.R
\name{sp.simplifyLen}
\alias{sp.simplifyLen}
\title{Simplify Polylines by length
\code{sp.simplifyLen}}
\usage{
sp.simplifyLen(sp, tol, threshold = tol/2)
}
\arguments{
\item{sp}{a SpatialLines or SpatialPoygons object}

\item{tol}{Tolerance}

\item{threshold}{Minimum length for a segment}
}
\description{
Simplify Polylines by length
\code{sp.simplifyLen}
}
