% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trianglulate.R
\name{sp.Tri2Shape}
\alias{sp.Tri2Shape}
\title{Generate triangle mesh domain
\code{sp.Tri2Shape}}
\usage{
sp.Tri2Shape(tri, dbf = NULL, crs = NA)
}
\arguments{
\item{tri}{triangulate}

\item{dbf}{attribute table, data.frame or matrix}

\item{crs}{Projection}
}
\value{
Coordinates of triangles centroids
}
\description{
Generate triangle mesh domain
\code{sp.Tri2Shape}
}
