% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{sp2PSLG}
\alias{sp2PSLG}
\title{Convert SpatialLines or SpatialPolygons a Planar Straight Line Graph object
\code{sp2PSLG}}
\usage{
sp2PSLG(sp)
}
\arguments{
\item{sp}{SpatialLines or SpatialPolygons}
}
\value{
pslg class
}
\description{
Convert SpatialLines or SpatialPolygons a Planar Straight Line Graph object
\code{sp2PSLG}
}
\examples{
if (requireNamespace("sf", quietly = TRUE) && requireNamespace("sp", quietly = TRUE)) {
  wkt <- "MULTIPOLYGON(((1 1,5 1,5 5,1 5,1 1),(2 2,2 3,3 3,3 2,2 2)))"
  sl <- sf::st_sf(geometry = sf::st_as_sfc(wkt))
  x <- sp2PSLG(sl)
}
}
