% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Misc.R
\name{sp2PSLG}
\alias{sp2PSLG}
\title{Convert SpatialLines or SpatialPolygons a Planar Straight Line Graph object
\code{sp2PSLG}}
\usage{
sp2PSLG(sp)
}
\arguments{
\item{sp}{SpatialLines or SpatialPolygons}
}
\value{
pslg class
}
\description{
Convert SpatialLines or SpatialPolygons a Planar Straight Line Graph object
\code{sp2PSLG}
}
\examples{
library(rgeos)
sl = readWKT("MULTIPOLYGON(((1 1,5 1,5 5,1 5,1 1),(2 2,2 3,3 3,3 2,2 2)),((6 3,9 2,9 4,6 3)))")
x = sp2PSLG(sl)
}
