% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_core.R
\name{sp2raster}
\alias{sp2raster}
\title{Deprecated: sp2raster}
\usage{
sp2raster(...)
}
\arguments{
\item{...}{Arguments passed to \code{vector_to_raster}}
}
\description{
This function is deprecated. Use \code{\link{vector_to_raster}} instead.
}
\keyword{internal}
