% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{sp2raster}
\alias{sp2raster}
\title{SpatialData to Raster
\code{sp2raster}}
\usage{
sp2raster(sp, mask = get("MASK", envir = .shud), ngrids = 200,
  resolution = NULL, field = 1)
}
\arguments{
\item{sp}{SpatialPolygon}

\item{mask}{Raster mask of mesh domain.}

\item{ngrids}{Number of grid along x direction.}

\item{resolution}{Resolution, defaul = NULL, resolution = extent / ngrids}

\item{field}{Index of field}
}
\value{
Raster map
}
\description{
SpatialData to Raster
\code{sp2raster}
}
