% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeshDomain.R
\name{Tri2Centroid}
\alias{Tri2Centroid}
\title{Centroids of the triangulation
\code{Tri2Centroid}}
\usage{
Tri2Centroid(tri)
}
\arguments{
\item{tri}{Triangles defination}
}
\value{
Centroids of the triangles, m x 2;
}
\description{
Centroids of the triangulation
\code{Tri2Centroid}
}
