% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelClasses.R
\docType{class}
\name{Untructure Domain-class}
\alias{Untructure Domain-class}
\alias{SHUD.MESH}
\title{Class of SHUD.MESH
\code{SHUD.MESH}}
\value{
Class of SHUD.MESH
}
\description{
Class of SHUD.MESH
\code{SHUD.MESH}
}
\section{Slots}{

\describe{
\item{\code{mesh}}{data.frame or matrix}

\item{\code{point}}{data.frame or matrix}
}}

