% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{VaporPressure_Act}
\alias{VaporPressure_Act}
\title{Actual vapor pressure 
\code{VaporPressure_Act}}
\usage{
VaporPressure_Act(esat, rh)
}
\arguments{
\item{esat}{Saturated vapor pressure}

\item{rh}{Relative Humidity [-] 0~1}
}
\description{
Actual vapor pressure 
\code{VaporPressure_Act}
}
