% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{voronoipolygons}
\alias{voronoipolygons}
\title{Generate Thiessen polygons from point data
\code{voronoipolygons}}
\usage{
voronoipolygons(x = NULL, pts = NULL, rw = NULL, crs = NULL)
}
\arguments{
\item{x}{\code{sf} point object in projected coordinates}

\item{pts}{Coordinates (x,y) of the points}

\item{rw}{The coordinates of the corners of the rectangular window enclosing the triangulation, in the order (xmin, xmax, ymin, ymax).  (default = NULL)}

\item{crs}{projection parameters (default = NULL)}
}
\value{
Legacy \code{SpatialPolygons*} object for compatibility.
}
\description{
Generate Thiessen polygons from point data
\code{voronoipolygons}
}
\examples{
library(rSHUD)
set.seed(1)
n = 10
xx = rnorm(n)
yy = rnorm(n)
x = sf::st_as_sf(data.frame(x = xx, y = yy), coords = c("x", "y"))
vx = voronoipolygons(x)
plot(sf::st_geometry(sf::st_as_sf(vx)), axes = TRUE)
plot(sf::st_geometry(x), add = TRUE, col = 2)
#' ====END====

x = 1:5
y = 1:5
xy = expand.grid(x, y)
vx = voronoipolygons(pts = xy)
plot(sf::st_geometry(sf::st_as_sf(vx)), axes = TRUE)
points(xy)
#' ====END====

library(sf)
set.seed(2)
n = 10
xx = rnorm(n)
yy = rnorm(n)
x = sf::st_as_sf(data.frame(x = xx, y = yy), coords = c("x", "y"))
y = sf::st_as_sf(data.frame(x = xx + 2, y = yy + 2), coords = c("x", "y"))
e1 = sf::st_bbox(y)
e2 = sf::st_bbox(x)
rw = c(xmin = min(e1["xmin"], e2["xmin"]),
       xmax = max(e1["xmax"], e2["xmax"]),
       ymin = min(e1["ymin"], e2["ymin"]),
       ymax = max(e1["ymax"], e2["ymax"])) + c(xmin = -1, xmax = 1, ymin = -1, ymax = 1)
vx = voronoipolygons(x = x, rw = rw) 
plot(sf::st_geometry(sf::st_as_sf(vx)))
plot(sf::st_geometry(x), add = TRUE, col = 2)
plot(sf::st_geometry(y), add = TRUE, col = 3)
}
