% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PET.R
\name{WindProfile}
\alias{WindProfile}
\title{Calculate windspeed at Zx heights.
\code{WindProfile}}
\usage{
WindProfile(zx, Um, zm = 10, d = 0.12, roughness = 0.012)
}
\arguments{
\item{zx}{Desired height [m]}

\item{Um}{Windspeed at measured height [m s-1]}

\item{zm}{Measured height [m s-1]}

\item{d}{Zero plane [m]}

\item{roughness}{roughness [m]}
}
\value{
windspeed at any desired heights.
}
\description{
Calculate windspeed at Zx heights.
\code{WindProfile}
}
