% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{write.config}
\alias{write.config}
\title{Deprecated wrapper for \code{write.config()}}
\usage{
write.config(x, file)
}
\description{
Legacy compatibility wrapper for \code{\link{write_config}}.
}
