% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{write.config}
\alias{write.config}
\title{Write SHUD configuration files (.para, .calib, etc.)}
\usage{
write.config(x, file)
}
\arguments{
\item{x}{SHUD model configure parameter or calibration}

\item{file}{file name}
}
\description{
Write SHUD configuration files (.para, .calib, etc.)
}
