% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_wrappers.R
\name{write.forc}
\alias{write.forc}
\title{Deprecated wrapper for \code{write.forc()}}
\usage{
write.forc(x, file, path = "", startdate = "20000101")
}
\description{
Legacy compatibility wrapper for \code{\link{write_forc}}.
}
