% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{write.forc}
\alias{write.forc}
\title{Write SHUD .forc file
\code{write.forc}}
\usage{
write.forc(x, file, path = "", startdate = "20000101")
}
\arguments{
\item{x}{data.frame of the forcing sites information. c(id, lon, lat, x, y, z, filename)}

\item{file}{file name}

\item{path}{Common path of the files.}

\item{startdate}{Start Date. Character. e.g. 20000101}
}
\description{
Write SHUD .forc file
\code{write.forc}
}
