% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{write.forc}
\alias{write.forc}
\title{Write SHUD .forc file
\code{write.forc}}
\usage{
write.forc(fns, path = "", startdate = "20000101", file, backup = TRUE)
}
\arguments{
\item{fns}{filenames; Vector of character,}

\item{path}{Common path of the files.}

\item{startdate}{Start Date. Character. e.g. 20000101}

\item{file}{file name}

\item{backup}{TRUE/FALSE, if backup the existing file}
}
\description{
Write SHUD .forc file
\code{write.forc}
}
