% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{write.ic}
\alias{write.ic}
\title{Write SHUD .ic file
\code{writeinit}}
\usage{
write.ic(x, file, backup = TRUE)
}
\arguments{
\item{x}{Initial condition, list()}

\item{file}{file name}

\item{backup}{TRUE/FALSE, if backup the existing file}
}
\description{
Write SHUD .ic file
\code{writeinit}
}
