% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{write.mesh}
\alias{write.mesh}
\title{Write  .mesh file
\code{write.mesh}}
\usage{
write.mesh(pm, file)
}
\arguments{
\item{pm}{SHUD.MESH class}

\item{file}{file name}
}
\description{
Write  .mesh file
\code{write.mesh}
}
