% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{write.riv}
\alias{write.riv}
\title{Write SHUD .mesh file
\code{writemesh}}
\usage{
write.riv(pr, file, backup = TRUE)
}
\arguments{
\item{pr}{SHUDmesh class}

\item{file}{file name}

\item{backup}{TRUE/FALSE, if backup the existing file}
}
\description{
Write SHUD .mesh file
\code{writemesh}
}
