% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInput.R
\name{write.tsd}
\alias{write.tsd}
\title{Write xts data out into file
\code{write.tsd}}
\usage{
write.tsd(x, file, append = F, quite = F, header = NULL,
  backup = TRUE)
}
\arguments{
\item{x}{xts data}

\item{file}{file name}

\item{append}{whether append}

\item{quite}{TRUE/FALSE, if quiet mode}

\item{header}{Header of the file. Default is the dimension of data.frame.}

\item{backup}{TRUE/FALSE, if backup the existing file}
}
\description{
Write xts data out into file
\code{write.tsd}
}
