% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{writeshape}
\alias{writeshape}
\title{Write ESRI shapefile out
\code{writeshape}}
\usage{
writeshape(shp, file = NULL, crs = raster::crs(shp))
}
\arguments{
\item{shp}{Spatial file}

\item{file}{file path, without '.shp'.}

\item{crs}{projection}
}
\description{
Write ESRI shapefile out
\code{writeshape}
}
\examples{
library(sp)
library(rgeos)
library(rgdal)
sp1 = readWKT("POLYGON((0 0, 0 1, 1 1, 1 0, 0 0))")
raster::crs(sp1) =sp::CRS("+init=epsg:4326")
writeshape(sp1, file=file.path(tempdir(), 'sp1'))
sp2=readOGR(file.path(tempdir(), 'sp1.shp'))
plot(sp2)
}
