% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{xy2shp}
\alias{xy2shp}
\title{Generate a shapefile from coordinates.
\code{xy2shp}}
\usage{
xy2shp(xy, df = data.frame(ID = 1:nrow(xy), xy), crs = NULL)
}
\arguments{
\item{xy}{matrix}

\item{df}{attribute table}

\item{crs}{projection parameters}
}
\value{
SpatialPointsDataFrame
}
\description{
Generate a shapefile from coordinates.
\code{xy2shp}
}
