% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_GIS.R
\name{xy2shp}
\alias{xy2shp}
\title{Generate a shapefile from coordinates.
\code{xy2shp}}
\usage{
xy2shp(xy, df = NULL, crs = NULL, shape = "points")
}
\arguments{
\item{xy}{matrix}

\item{df}{attribute table}

\item{crs}{projection parameters}

\item{shape}{Shape of the result in points, lines or polygons}
}
\value{
SpatialPointsDataFrame, SpatialLinesDataFrame, or SpatialPolygonsDataFrame
}
\description{
Generate a shapefile from coordinates.
\code{xy2shp}
}
\examples{
library(raster)
xy=list(cbind(c(0, 2, 1), c(0, 0, 2)),  cbind(c(0, 2, 1), c(0, 0, 2))+2)
sp1 = xy2shp(xy=xy, shape = 'polygon')
raster::plot(sp1, axes=TRUE, col='gray')

sp2 = xy2shp(xy=xy, shape = 'lines')
raster::plot(sp2, add=TRUE, lty=2, lwd=3,col='red')
sp3 = xy2shp(xy=xy, shape = 'POINTS')
raster::plot(sp3, add=TRUE, pch=1, cex=2)

}
